Crate libp2p::allow_block_list

A libp2p module for managing allow and blocks lists to peers.

Allow list example

#[derive(NetworkBehaviour)]
struct MyBehaviour {
   allowed_peers: allow_block_list::Behaviour<AllowedPeers>,
}

let behaviour = MyBehaviour {
    allowed_peers: allow_block_list::Behaviour::default()
};

Block list example

#[derive(NetworkBehaviour)]
struct MyBehaviour {
   blocked_peers: allow_block_list::Behaviour<BlockedPeers>,
}

let behaviour = MyBehaviour {
    blocked_peers: allow_block_list::Behaviour::default()
};

Structs

AllowedPeers

The list of explicitly allowed peers.

Behaviour

A NetworkBehaviour that can act as an allow or block list.

Blocked

A connection to this peer was explicitly blocked and was thus denied.

BlockedPeers

The list of explicitly blocked peers.

NotAllowed

A connection to this peer is not explicitly allowed and was thus denied.