Module libp2p::autonat::v1

Implementation of the AutoNAT protocol.

Eventual Deprecation

This version of the protocol will eventually be deprecated in favor of v2. We recommend using v2 for new projects.

Structs

Behaviour

NetworkBehaviour for AutoNAT.

Config

Config for the Behaviour.

ProbeId

Unique identifier for a probe.

Enums

Event

Event produced by Behaviour.

InboundFailure

Possible failures occurring in the context of receiving an inbound request and sending a response.

InboundProbeError

Inbound probe failed.

InboundProbeEvent

NatStatus

Assumed NAT status.

OutboundFailure

Possible failures occurring in the context of sending an outbound request and receiving the response.

OutboundProbeError

Outbound probe failed or was aborted.

OutboundProbeEvent

ResponseError

Constants

DEFAULT_PROTOCOL_NAME

The protocol name used for negotiating with multistream-select.