Module libp2p::bandwidth

Structs

BandwidthConnecLogging
Wraps around an AsyncRead + AsyncWrite and logs the bandwidth that goes through it.

BandwidthFuture
Wraps around a Future that produces a connection. Wraps the connection around a bandwidth counter.

BandwidthListener
Wraps around a Stream that produces connections. Wraps each connection around a bandwidth counter.

BandwidthLogging
Wraps around a Transport and counts the number of bytes that go through all the opened connections.

BandwidthSinks
Allows obtaining the average bandwidth of the connections created from a BandwidthLogging.