Module libp2p::core

Transports, upgrades, multiplexing and node handling of libp2p.

The main concepts of libp2p-core are:

A [PeerId] is a unique global identifier for a node on the network. Each node must have a different PeerId. Normally, a PeerId is the hash of the public key used to negotiate encryption on the communication channel, thereby guaranteeing that they cannot be spoofed.

The [Transport] trait defines how to reach a remote node or listen for incoming remote connections. See the transport module.

The [StreamMuxer] trait is implemented on structs that hold a connection to a remote and can subdivide this connection into multiple substreams. See the muxing module.

The [UpgradeInfo], [InboundUpgrade] and [OutboundUpgrade] traits define how to upgrade each individual substream to use a protocol. See the upgrade module.

Modules

either

identity
A node's network identity keys.

multiaddr

muxing
Muxing is the process of splitting a connection into multiple substreams.

nodes
Low-level networking primitives.

transport
Connection-oriented communication channels.

upgrade
Contains everything related to upgrading a connection or a substream to use a protocol.

Structs

PeerId
Identifier of a peer of the network.

Enums

ConnectedPoint
How we connected to a node.

Endpoint

Functions

address_translation
Perform IP address translation.