Enum libp2p::core::multiaddr::Protocol

pub enum Protocol<'a> {

    Dccp(u16),
    Dns(Cow<'a, str>),
    Dns4(Cow<'a, str>),
    Dns6(Cow<'a, str>),
    Dnsaddr(Cow<'a, str>),
    Http,
    Https,
    Ip4(Ipv4Addr),
    Ip6(Ipv6Addr),
    P2pWebRtcDirect,
    P2pWebRtcStar,
    P2pWebSocketStar,
    Memory(u64),
    Onion(Cow<'a, [u8; 10]>, u16),
    Onion3(Onion3Addr<'a>),
    P2p(Multihash<UInt<UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>, B0>, B0>, B0>>),
    P2pCircuit,
    Quic,
    Sctp(u16),
    Tcp(u16),
    Udp(u16),
    Udt,
    Unix(Cow<'a, str>),
    Utp,
    Ws(Cow<'a, str>),
    Wss(Cow<'a, str>),

}

Protocol describes all possible multiaddress protocols.

For Unix, Ws and Wss we use &str instead of Path to allow cross-platform usage of Protocol since encoding Paths to bytes is platform-specific. This means that the actual validation of paths needs to happen separately.

Variants

Dccp(u16)

Tuple Fields

0: u16

Dns(Cow<'a, str>)

Tuple Fields

0: Cow<'a, str>

Dns4(Cow<'a, str>)

Tuple Fields

0: Cow<'a, str>

Dns6(Cow<'a, str>)

Tuple Fields

0: Cow<'a, str>

Dnsaddr(Cow<'a, str>)

Tuple Fields

0: Cow<'a, str>

Http

Https

Ip4(Ipv4Addr)

Tuple Fields

0: Ipv4Addr

Ip6(Ipv6Addr)

Tuple Fields

0: Ipv6Addr

P2pWebRtcDirect

P2pWebRtcStar

P2pWebSocketStar

Memory(u64)

Tuple Fields

0: u64

Contains the “port” to contact. Similar to TCP or UDP, 0 means “assign me a port”.

Onion(Cow<'a, [u8; 10]>, u16)

Tuple Fields

0: Cow<'a, [u8; 10]>

1: u16

Onion3(Onion3Addr<'a>)

Tuple Fields

0: Onion3Addr<'a>

P2p(Multihash<UInt<UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>, B0>, B0>, B0>>)

Tuple Fields

0: Multihash<UInt<UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>, B0>, B0>, B0>>

P2pCircuit

Quic

Sctp(u16)

Tuple Fields

0: u16

Tcp(u16)

Tuple Fields

0: u16

Udp(u16)

Tuple Fields

0: u16

Udt

Unix(Cow<'a, str>)

Tuple Fields

0: Cow<'a, str>

Utp

Ws(Cow<'a, str>)

Tuple Fields

0: Cow<'a, str>

Wss(Cow<'a, str>)

Tuple Fields

0: Cow<'a, str>

Implementations

impl<'a> Protocol<'a>

pub fn from_str_parts<I>(iter: I) -> Result<Protocol<'a>, Error> where
I: Iterator<Item = &'a str>,

Parse a protocol value from the given iterator of string slices.

The parsing only consumes the minimum amount of string slices necessary to produce a well-formed protocol. The same iterator can thus be used to parse a sequence of protocols in succession. It is up to client code to check that iteration has finished whenever appropriate.

pub fn from_bytes(input: &'a [u8]) -> Result<(Protocol<'a>, &'a [u8]), Error>

Parse a single Protocol value from its byte slice representation, returning the protocol as well as the remaining byte slice.

pub fn write_bytes<W>(&self, w: &mut W) -> Result<(), Error> where
W: Write,

Encode this protocol by writing its binary representation into the given Write impl.

pub fn acquire<'b>(self) -> Protocol<'b>

Turn this Protocol into one that owns its data, thus being valid for any lifetime.