Crate libp2p::core::multiaddr::multihash

Bare-minimum multihash data structure.

This crate defines a no_std compatible data structures for representing a Multihash.

It does not offer any hashing, instead you are encouraged to either do the hashing yourself. Alternatively, you can use an existing code table or make your own code table.

The multihash-codetable crate defines a set of hashes to get started quickly. To make your own codetable, use the multihash-derive crate.

The arb feature flag enables the quickcheck arbitrary implementation for property based testing.

For serializing the multihash there is support for Serde via the serde-codec feature and the SCALE Codec via the scale-codec feature.

Structs

Error

Opaque error struct for operations involving a Multihash.

Multihash

A Multihash instance that only supports the basic functionality and no hashing.

Type Definitions

MultihashGenericDeprecated

Deprecated type-alias for the Multihash type.

ResultDeprecated

Multihash result.