Trait libp2p::core::multiaddr::multihash::Digest

pub trait Digest<S>: 'static + AsRef<[u8]> + AsMut<[u8]> + From<GenericArray<u8, S>> + Into<GenericArray<u8, S>> + Clone + Hash + Debug + Default + Eq + Send + Sync where
    S: Size, {
    pub fn size(&self) -> u8 { ... }

    pub fn wrap(digest: &[u8]) -> Result<Self, Error> { ... }

    pub fn from_reader<R>(r: R) -> Result<Self, Error>
    where
        R: Read,
    { ... }
}

Stack allocated digest trait.

Provided methods

pub fn size(&self) -> u8

Size of the digest.

pub fn wrap(digest: &[u8]) -> Result<Self, Error>

Wraps the digest bytes.

pub fn from_reader<R>(r: R) -> Result<Self, Error> where
R: Read,

Reads a multihash digest from a byte stream that contains the digest prefixed with the size.

The byte stream must not contain the code as prefix.

Loading content...