Trait libp2p::core::multiaddr::multihash::Multihasher

pub trait Multihasher<T> where
    T: TryFrom<u64> + Copy, {
    const CODE: T;

    fn digest(data: &[u8]) -> MultihashGeneric<T>;
}

The Multihasher trait specifies an interface common for all multihash functions that does not require allocating a Box<dyn MultihashDigest<T>>.

Associated Constants

const CODE: T

The multihash code.

Loading content...
Required methods

fn digest(data: &[u8]) -> MultihashGeneric<T>

Hash some input and return the digest.

Loading content...