Function libp2p::core::muxing::inbound_from_ref_and_wrap

pub fn inbound_from_ref_and_wrap<P>(
    muxer: P
) -> impl Future<Item = SubstreamRef<P>, Error = Error> where
    P: Deref + Clone,
    <P as Deref>::Target: StreamMuxer, 

Polls for an inbound from the muxer but wraps the output in an object that implements Read/Write/AsyncRead/AsyncWrite.