Struct libp2p::core::muxing::SingletonMuxer

pub struct SingletonMuxer<TSocket> { /* fields omitted */ }

Implementation of StreamMuxer that allows only one substream on top of a connection, yielding the connection itself.

Applying this muxer on a connection doesn't read or write any data on the connection itself. Most notably, no protocol is negotiated.

Methods

impl<TSocket> SingletonMuxer<TSocket>

pub fn new(inner: TSocket, endpoint: Endpoint) -> SingletonMuxer<TSocket>

Creates a new SingletonMuxer.

If endpoint is Dialer, then only one outbound substream will be permitted. If endpoint is Listener, then only one inbound substream will be permitted.