Enum libp2p::core::network::peer::Peer

pub enum Peer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId> where
    THandler: IntoConnectionHandler<TConnInfo>,
    TTrans: Transport,  {
    Connected(ConnectedPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>),
    Dialing(DialingPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>),
    Disconnected(DisconnectedPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>),
    Local,
}

The possible representations of a peer in a Network, as seen by the local node.

Note: In any state there may always be a pending incoming connection attempt from the peer, however, the remote identity of a peer is only known once a connection is fully established.

Variants

Connected(ConnectedPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>)

At least one established connection exists to the peer.

Dialing(DialingPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>)

There is an ongoing dialing (i.e. outgoing connection) attempt to the peer. There may already be other established connections to the peer.

Disconnected(DisconnectedPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>)

There exists no established connection to the peer and there is currently no ongoing dialing (i.e. outgoing connection) attempt in progress.

Local

The peer represents the local node.

Implementations

impl<'a, TTrans, TMuxer, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId> Peer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId> where
TConnInfo: Debug + ConnectionInfo<PeerId = TPeerId> + Send + 'static,
THandler: IntoConnectionHandler<TConnInfo> + Send + 'static,
TInEvent: Send + 'static,
TMuxer: StreamMuxer + Send + Sync + 'static,
TOutEvent: Send + 'static,
TPeerId: Eq + Hash + Clone + Send + 'static,
TTrans: Transport<Output = (TConnInfo, TMuxer)> + Clone,
<TTrans as Transport>::Error: Send,
<TTrans as Transport>::Error: 'static,
<TTrans as Transport>::Dial: Send,
<TTrans as Transport>::Dial: 'static,
<TMuxer as StreamMuxer>::OutboundSubstream: Send,
<THandler as IntoConnectionHandler<TConnInfo>>::Handler: ConnectionHandler,
<THandler as IntoConnectionHandler<TConnInfo>>::Handler: Send,
<<THandler as IntoConnectionHandler<TConnInfo>>::Handler as ConnectionHandler>::OutboundOpenInfo: Send,
<<THandler as IntoConnectionHandler<TConnInfo>>::Handler as ConnectionHandler>::Error: Error,
<<THandler as IntoConnectionHandler<TConnInfo>>::Handler as ConnectionHandler>::Error: Send,
<<THandler as IntoConnectionHandler<TConnInfo>>::Handler as ConnectionHandler>::Error: 'static,
<<THandler as IntoConnectionHandler<TConnInfo>>::Handler as ConnectionHandler>::Substream == SubstreamRef<Arc<TMuxer>>,
<<THandler as IntoConnectionHandler<TConnInfo>>::Handler as ConnectionHandler>::InEvent == TInEvent,
<<THandler as IntoConnectionHandler<TConnInfo>>::Handler as ConnectionHandler>::OutEvent == TOutEvent,

pub fn is_connected(&self) -> bool

Checks whether the peer is currently connected.

Returns true iff Peer::into_connected returns Some.

pub fn is_dialing(&self) -> bool

Checks whether the peer is currently being dialed.

Returns true iff Peer::into_dialing returns Some.

pub fn is_disconnected(&self) -> bool

Checks whether the peer is currently disconnected.

Returns true iff Peer::into_disconnected returns Some.

pub fn dial<I>(
self,
address: Multiaddr,
remaining: I,
handler: THandler
) -> Result<(ConnectionId, DialingPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>), ConnectionLimit> where
I: IntoIterator<Item = Multiaddr>,

Initiates a new dialing attempt to this peer using the given addresses.

The connection ID of the first connection attempt, i.e. to address, is returned, together with a DialingPeer for further use. The remaining addresses are tried in order in subsequent connection attempts in the context of the same dialing attempt, if the connection attempt to the first address fails.

pub fn into_connected(
self
) -> Option<ConnectedPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>>

Converts the peer into a ConnectedPeer, if an established connection exists.

Succeeds if the there is at least one established connection to the peer.

pub fn into_dialing(
self
) -> Option<DialingPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>>

Converts the peer into a DialingPeer, if a dialing attempt exists.

Succeeds if the there is at least one pending outgoing connection to the peer.

pub fn into_disconnected(
self
) -> Option<DisconnectedPeer<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>>

Converts the peer into a DisconnectedPeer, if neither an established connection nor a dialing attempt exists.