Struct libp2p::core::network::peer::ConnectedPeer

pub struct ConnectedPeer<'a, TTrans, TInEvent, TOutEvent, THandler> where
    THandler: IntoConnectionHandler,
    TTrans: Transport,  { /* fields omitted */ }

The representation of a peer in a Network to whom at least one established connection exists. There may also be additional ongoing dialing attempts to the peer.

Implementations

impl<'a, TTrans, TInEvent, TOutEvent, THandler> ConnectedPeer<'a, TTrans, TInEvent, TOutEvent, THandler> where
THandler: IntoConnectionHandler,
TTrans: Transport,

pub fn id(&self) -> &PeerId

pub fn into_peer(self) -> Peer<'a, TTrans, TInEvent, TOutEvent, THandler>

Returns the ConnectedPeer into a Peer.

pub fn connection(
&mut self,
id: ConnectionId
) -> Option<EstablishedConnection<'_, TInEvent>>

Obtains an established connection to the peer by ID.

pub fn num_connections(&self) -> u32

The number of established connections to the peer.

pub fn is_dialing(&self) -> bool

Checks whether there is an ongoing dialing attempt to the peer.

Returns true iff ConnectedPeer::into_dialing returns Some.

pub fn into_dialing(
self
) -> Option<DialingPeer<'a, TTrans, TInEvent, TOutEvent, THandler>>

Converts this peer into a DialingPeer, if there is an ongoing dialing attempt, None otherwise.

pub fn connections(
&mut self
) -> EstablishedConnectionIter<'_, impl Iterator<Item = ConnectionId>, TInEvent, TOutEvent, THandler, <TTrans as Transport>::Error, <<THandler as IntoConnectionHandler>::Handler as ConnectionHandler>::Error>

Gets an iterator over all established connections to the peer.

pub fn some_connection(&mut self) -> EstablishedConnection<'_, TInEvent>

Obtains some established connection to the peer.

pub fn disconnect(
self
) -> DisconnectedPeer<'a, TTrans, TInEvent, TOutEvent, THandler>

Disconnects from the peer, closing all connections.