Module libp2p::core::nodes

Low-level networking primitives.

Contains structs that are aiming at providing very precise control over what happens over the network.

The more complete and highest-level struct is the Network. The Network directly or indirectly uses all the other structs of this module.

Modules

collection

handled_node

listeners
Manage listening on multiple multiaddresses at once.

network

node

tasks
Management of tasks handling nodes.