Enum libp2p::core::nodes::listeners::ListenersEvent

pub enum ListenersEvent<TTrans> where
    TTrans: Transport,  {
    NewAddress {
        listen_addr: Multiaddr,
    },
    AddressExpired {
        listen_addr: Multiaddr,
    },
    Incoming {
        upgrade: <TTrans as Transport>::ListenerUpgrade,
        listen_addr: Multiaddr,
        send_back_addr: Multiaddr,
    },
    Closed {
        listener: <TTrans as Transport>::Listener,
        result: Result<(), <<TTrans as Transport>::Listener as Stream>::Error>,
    },
}

Event that can happen on the ListenersStream.

Variants
NewAddress

A new address is being listened on.

Fields of NewAddress

listen_addr: Multiaddr

The new address that is being listened on.

AddressExpired

An address is no longer being listened on.

Fields of AddressExpired

listen_addr: Multiaddr

The new address that is being listened on.

Incoming

A connection is incoming on one of the listeners.

Fields of Incoming

upgrade: <TTrans as Transport>::ListenerUpgrade

The produced upgrade.

listen_addr: Multiaddr

Address of the listener which received the connection.

send_back_addr: Multiaddr

Address used to send back data to the incoming client.

Closed

A listener has closed, either gracefully or with an error.

Fields of Closed

listener: <TTrans as Transport>::Listener

The listener that closed.

result: Result<(), <<TTrans as Transport>::Listener as Stream>::Error>

The error that happened. Ok if gracefully closed.