Struct libp2p::core::nodes::raw_swarm::RawSwarm

pub struct RawSwarm<TTrans, TInEvent, TOutEvent, THandler, THandlerErr, TConnInfo = PeerId, TPeerId = PeerId> where
    TTrans: Transport,  { /* fields omitted */ }

Implementation of Stream that handles the nodes.

Methods

impl<TTrans, TInEvent, TOutEvent, TMuxer, THandler, THandlerErr, TConnInfo, TPeerId> RawSwarm<TTrans, TInEvent, TOutEvent, THandler, THandlerErr, TConnInfo, TPeerId> where
TConnInfo: Debug + ConnectionInfo<PeerId = TPeerId> + Send + 'static,
THandler: IntoNodeHandler<(TConnInfo, ConnectedPoint)> + Send + 'static,
THandlerErr: Error + Send + 'static,
TMuxer: StreamMuxer,
TPeerId: Eq + Hash + Clone,
TTrans: Transport + Clone,
<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler: NodeHandler,
<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler: Send,
<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler: 'static,
<<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler as NodeHandler>::OutboundOpenInfo: Send,
<<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler as NodeHandler>::OutboundOpenInfo: 'static,
<<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler as NodeHandler>::Substream == SubstreamRef<Arc<TMuxer>>,
<<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler as NodeHandler>::InEvent == TInEvent,
<<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler as NodeHandler>::OutEvent == TOutEvent,
<<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler as NodeHandler>::Error == THandlerErr,

pub fn new(
transport: TTrans,
local_peer_id: TPeerId
) -> RawSwarm<TTrans, TInEvent, TOutEvent, THandler, THandlerErr, TConnInfo, TPeerId>

Creates a new node events stream.

pub fn new_with_incoming_limit(
transport: TTrans,
local_peer_id: TPeerId,
incoming_limit: Option<u32>
) -> RawSwarm<TTrans, TInEvent, TOutEvent, THandler, THandlerErr, TConnInfo, TPeerId>

Creates a new node event stream with incoming connections limit.

pub fn transport(&self) -> &TTrans

Returns the transport passed when building this object.

pub fn listen_on(
&mut self,
addr: Multiaddr
) -> Result<(), TransportError<<TTrans as Transport>::Error>>

Start listening on the given multiaddress.

pub fn listen_addrs(&self) -> impl Iterator<Item = &Multiaddr>

Returns an iterator that produces the list of addresses we are listening on.

pub fn incoming_limit(&self) -> Option<u32>

Returns limit on incoming connections.

pub fn nat_traversal<'a>(
&'a self,
observed_addr: &'a Multiaddr
) -> impl Iterator<Item = Multiaddr> + 'a where
THandler: 'a,
TMuxer: 'a,

Call this function in order to know which address remotes should dial in order to access your local node.

observed_addr should be an address a remote observes you as, which can be obtained for example with the identify protocol.

For each listener, calls nat_traversal with the observed address and returns the outcome.

pub fn local_peer_id(&self) -> &TPeerId

Returns the peer id of the local node.

This is the same value as was passed to new().

pub fn dial(
&mut self,
addr: Multiaddr,
handler: THandler
) -> Result<(), TransportError<<TTrans as Transport>::Error>> where
TConnInfo: Send + 'static,
TInEvent: Send + 'static,
TMuxer: StreamMuxer + Send + Sync + 'static,
TOutEvent: Send + 'static,
TPeerId: Send + 'static,
TTrans: Transport<Output = (TConnInfo, TMuxer)>,
<TTrans as Transport>::Error: Send,
<TTrans as Transport>::Error: 'static,
<TTrans as Transport>::Dial: Send,
<TTrans as Transport>::Dial: 'static,
<TMuxer as StreamMuxer>::OutboundSubstream: Send,
<TMuxer as StreamMuxer>::Substream: Send,

Dials a multiaddress without knowing the peer ID we're going to obtain.

The second parameter is the handler to use if we manage to reach a node.

pub fn num_incoming_negotiated(&self) -> usize

Deprecated:
Use incoming_negotiated().count() instead

Returns the number of incoming connections that are currently in the process of being negotiated.

We don't know anything about these connections yet, so all we can do is know how many of them we have.

pub fn incoming_negotiated(&self) -> impl Iterator<Item = IncomingInfo>

Returns the list of incoming connections that are currently in the process of being negotiated. We don't know the PeerId of these nodes yet.

pub fn broadcast_event(&mut self, event: &TInEvent) where
TInEvent: Clone,

Sends an event to all nodes.

pub fn connected_peers(&self) -> impl Iterator<Item = &TPeerId>

Returns a list of all the peers we are currently connected to.

Calling peer() with each PeerId is guaranteed to produce a PeerConnected.

pub fn pending_connection_peers(&self) -> impl Iterator<Item = &TPeerId>

Returns a list of all the nodes we are currently trying to reach.

Calling peer() with each PeerId is guaranteed to produce a PeerPendingConnect

pub fn unknown_dials(&self) -> impl Iterator<Item = &Multiaddr>

Returns the list of addresses we're currently dialing without knowing the PeerId of.

pub fn peer(
&mut self,
peer_id: TPeerId
) -> Peer<TTrans, TInEvent, TOutEvent, THandler, THandlerErr, TConnInfo, TPeerId>

Grants access to a struct that represents a peer.

pub fn poll(
&mut self
) -> Async<RawSwarmEvent<TTrans, TInEvent, TOutEvent, THandler, THandlerErr, TConnInfo, TPeerId>> where
TConnInfo: Clone,
THandler: IntoNodeHandler<(TConnInfo, ConnectedPoint)> + Send + 'static,
THandlerErr: Error + Send + 'static,
TInEvent: Send + 'static,
TMuxer: StreamMuxer + Send + Sync + 'static,
TOutEvent: Send + 'static,
TPeerId: AsRef<[u8]> + Send + 'static,
TTrans: Transport<Output = (TConnInfo, TMuxer)>,
<TTrans as Transport>::Error: Send,
<TTrans as Transport>::Error: 'static,
<TTrans as Transport>::Dial: Send,
<TTrans as Transport>::Dial: 'static,
<TTrans as Transport>::ListenerUpgrade: Send,
<TTrans as Transport>::ListenerUpgrade: 'static,
<TMuxer as StreamMuxer>::OutboundSubstream: Send,
<TMuxer as StreamMuxer>::Substream: Send,
<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler: NodeHandler,
<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler: Send,
<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler: 'static,
<<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler as NodeHandler>::OutboundOpenInfo: Send,
<<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler as NodeHandler>::OutboundOpenInfo: 'static,
<<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler as NodeHandler>::Substream == SubstreamRef<Arc<TMuxer>>,
<<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler as NodeHandler>::InEvent == TInEvent,
<<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler as NodeHandler>::OutEvent == TOutEvent,
<<THandler as IntoNodeHandler<(TConnInfo, ConnectedPoint)>>::Handler as NodeHandler>::Error == THandlerErr,

Provides an API similar to Stream, except that it cannot error.