Module libp2p::core::nodes::tasks

Management of tasks handling nodes.

The core type is a [task::Task], which implements [futures::Future] and connects and handles a node. A task receives and sends messages ([tasks::FromTaskMessage], [tasks::ToTaskMessage]) to the outside.

A set of tasks is managed by a [Manager] which creates tasks when a node should be connected to (cf. [Manager::add_reach_attempt]) or an existing connection to a node should be driven forward (cf. [Manager::add_connection]). Tasks can be referred to by [TaskId] and messages can be sent to individual tasks or all (cf. [Manager::start_broadcast]). Messages produces by tasks can be retrieved by polling the manager (cf. [Manager::poll]).

Structs

ClosedTask
Task after it has been closed.

Manager
Implementation of [Stream] that handles a collection of nodes.

TaskEntry
Access to a task in the collection.

TaskId
Task identifier.

Enums

Error
Error that can happen in a task.

Event
Event produced by the [Manager].