Struct libp2p::core::PeerRecord

pub struct PeerRecord { /* private fields */ }

Represents a peer routing record.

Peer records are designed to be distributable and carry a signature by being wrapped in a signed envelope. For more information see RFC0003 of the libp2p specifications: https://github.com/libp2p/specs/blob/master/RFC/0003-routing-records.md

Implementations

impl PeerRecord

pub fn from_signed_envelope( envelope: SignedEnvelope ) -> Result<PeerRecord, FromEnvelopeError>

Attempt to re-construct a PeerRecord from a SignedEnvelope.

If this function succeeds, the SignedEnvelope contained a peer record with a valid signature and can hence be considered authenticated.

pub fn new( key: &Keypair, addresses: Vec<Multiaddr> ) -> Result<PeerRecord, SigningError>

Construct a new PeerRecord by authenticating the provided addresses with the given key.

This is the same key that is used for authenticating every libp2p connection of your application, i.e. what you use when setting up your crate::transport::Transport.

pub fn to_signed_envelope(&self) -> SignedEnvelope

pub fn into_signed_envelope(self) -> SignedEnvelope

pub fn peer_id(&self) -> PeerId

pub fn seq(&self) -> u64

pub fn addresses(&self) -> &[Multiaddr]