Trait libp2p::core::OutboundUpgrade

pub trait OutboundUpgrade<C>: UpgradeInfo where
    <Self::Future as Future>::Output == Result<Self::Output, Self::Error>, {
    type Output;
    type Error;
    type Future: Future;
    fn upgrade_outbound(self, socket: C, info: Self::Info) -> Self::Future;
}

Possible upgrade on an outbound connection or substream.

Associated Types

type Output

Output after the upgrade has been successfully negotiated and the handshake performed.

type Error

Possible error during the handshake.

type Future: Future

Future that performs the handshake with the remote.

Loading content...
Required methods

fn upgrade_outbound(self, socket: C, info: Self::Info) -> Self::Future

After we have determined that the remote supports one of the protocols we support, this method is called to start the handshake.

The info is the identifier of the protocol, as produced by protocol_info.

Loading content...