Enum libp2p::core::transport::ListenerEvent

pub enum ListenerEvent<TUpgr, TErr> {
    NewAddress(Multiaddr),
    Upgrade {
        upgrade: TUpgr,
        local_addr: Multiaddr,
        remote_addr: Multiaddr,
    },
    AddressExpired(Multiaddr),
    Error(TErr),
}

Event produced by Transport::Listeners.

Transports are expected to produce Upgrade events only for listen addresses which have previously been announced via a NewAddress event and which have not been invalidated by an AddressExpired event yet.

Variants

NewAddress(Multiaddr)

The transport is listening on a new additional Multiaddr.

Upgrade

An upgrade, consisting of the upgrade future, the listener address and the remote address.

Fields of Upgrade

upgrade: TUpgr

The upgrade.

local_addr: Multiaddr

The local address which produced this upgrade.

remote_addr: Multiaddr

The remote address which produced this upgrade.

AddressExpired(Multiaddr)

A Multiaddr is no longer used for listening.

Error(TErr)

A non-fatal error has happened on the listener.

This event should be generated in order to notify the user that something wrong has happened. The listener, however, continues to run.

Implementations

impl<TUpgr, TErr> ListenerEvent<TUpgr, TErr>

pub fn map<U>(self, f: impl FnOnce(TUpgr) -> U) -> ListenerEvent<U, TErr>

In case this ListenerEvent is an upgrade, apply the given function to the upgrade and multiaddress and produce another listener event based the the function's result.

pub fn map_err<U>(self, f: impl FnOnce(TErr) -> U) -> ListenerEvent<TUpgr, U>

In case this ListenerEvent is an Error, apply the given function to the error and produce another listener event based on the function's result.

pub fn is_upgrade(&self) -> bool

Returns true if this is an Upgrade listener event.

pub fn into_upgrade(self) -> Option<(TUpgr, Multiaddr)>

Try to turn this listener event into upgrade parts.

Returns None if the event is not actually an upgrade, otherwise the upgrade and the remote address.

pub fn is_new_address(&self) -> bool

Returns true if this is a NewAddress listener event.

pub fn into_new_address(self) -> Option<Multiaddr>

Try to turn this listener event into the NewAddress part.

Returns None if the event is not actually a NewAddress, otherwise the address.

pub fn is_address_expired(&self) -> bool

Returns true if this is an AddressExpired listener event.

pub fn into_address_expired(self) -> Option<Multiaddr>

Try to turn this listener event into the AddressExpired part.

Returns None if the event is not actually a AddressExpired, otherwise the address.

pub fn is_error(&self) -> bool

Returns true if this is an Error listener event.

pub fn into_error(self) -> Option<TErr>

Try to turn this listener event into the Error part.

Returns None if the event is not actually a Error, otherwise the error.