Struct libp2p::core::transport::OptionalTransport

pub struct OptionalTransport<T>(/* private fields */);

Transport that is possibly disabled.

An OptionalTransport<T> is a wrapper around an Option<T>. If it is disabled (read: contains None), then any attempt to dial or listen will return MultiaddrNotSupported. If it is enabled (read: contains Some), then dialing and listening will be handled by the inner transport.

Implementations

impl<T> OptionalTransport<T>

pub fn some(inner: T) -> OptionalTransport<T>

Builds an OptionalTransport with the given transport in an enabled state.

pub fn none() -> OptionalTransport<T>

Builds a disabled OptionalTransport.