Module libp2p::core::transport::timeout

Transports with timeouts on the connection setup.

The connection setup includes all protocol upgrades applied on the underlying Transport.

Structs

Timeout

Wraps around a Future. Turns the error type from TimeoutError<Err> to TransportTimeoutError<Err>.

TimeoutListener

TransportTimeout

A TransportTimeout is a Transport that wraps another Transport and adds timeouts to all inbound and outbound connection attempts.

Enums

TransportTimeoutError

Error that can be produced by the TransportTimeout layer.