Trait libp2p::core::transport::Transport

pub trait Transport where
    <Self::Listener as Stream>::Item == ListenerEvent<Self::ListenerUpgrade>,
    <Self::Listener as Stream>::Error == Self::Error,
    <Self::ListenerUpgrade as Future>::Item == Self::Output,
    <Self::ListenerUpgrade as Future>::Error == Self::Error,
    <Self::Dial as Future>::Item == Self::Output,
    <Self::Dial as Future>::Error == Self::Error, {
    type Output;
    type Error: Error;
    type Listener: Stream;
    type ListenerUpgrade: Future;
    type Dial: Future;
    fn listen_on(
        self, 
        addr: Multiaddr
    ) -> Result<Self::Listener, TransportError<Self::Error>>;

    fn dial(
        self, 
        addr: Multiaddr
    ) -> Result<Self::Dial, TransportError<Self::Error>>;

    fn boxed(self) -> Boxed<Self::Output, Self::Error>
    where
        Self: Clone + Send + Sync + 'static,
        Self::Dial: Send,
        Self::Dial: 'static,
        Self::Listener: Send,
        Self::Listener: 'static,
        Self::ListenerUpgrade: Send,
        Self::ListenerUpgrade: 'static,
    { ... }

    fn map<F, O>(self, map: F) -> Map<Self, F>
    where
        F: FnOnce(Self::Output, ConnectedPoint) -> O + Clone,
    { ... }

    fn map_err<F, TNewErr>(self, map_err: F) -> MapErr<Self, F>
    where
        F: FnOnce(Self::Error) -> TNewErr + Clone,
    { ... }

    fn or_transport<T>(self, other: T) -> OrTransport<Self, T> { ... }

    fn with_upgrade<U, O, E>(self, upgrade: U) -> Upgrade<Self, U>
    where
        U: InboundUpgrade<Self::Output, Output = O, Error = E> + OutboundUpgrade<Self::Output, Output = O, Error = E>,
        Self::Output: AsyncRead,
        Self::Output: AsyncWrite,
    { ... }

    fn and_then<C, F, O>(self, upgrade: C) -> AndThen<Self, C>
    where
        C: FnOnce(Self::Output, ConnectedPoint) -> F + Clone,
        F: IntoFuture<Item = O>,
    { ... }

    fn with_timeout(self, timeout: Duration) -> TransportTimeout<Self> { ... }

    fn with_outbound_timeout(self, timeout: Duration) -> TransportTimeout<Self> { ... }

    fn with_inbound_timeout(self, timeout: Duration) -> TransportTimeout<Self> { ... }
}

A transport provides connection-oriented communication between two peers through ordered streams of data (i.e. connections).

Connections are established either by listening or dialing on a Transport. A peer that obtains a connection by listening is often referred to as the listener and the peer that initiated the connection through dialing as the dialer, in contrast to the traditional roles of server and client.

Most transports also provide a form of reliable delivery on the established connections but the precise semantics of these guarantees depend on the specific transport.

This trait is implemented for concrete connection-oriented transport protocols like TCP or Unix Domain Sockets, but also on wrappers that add additional functionality to the dialing or listening process (e.g. name resolution via the DNS).

Additional protocols can be layered on top of the connections established by a Transport through an upgrade mechanism that is initiated via with_upgrade and optionally followed by further upgrades through chaining calls to with_upgrade and and_then. Thereby every upgrade yields a new Transport whose connection setup incorporates all earlier upgrades followed by the new upgrade, i.e. the order of the upgrades is significant.

Note: The methods of this trait use self and not &self or &mut self. In other words, listening or dialing consumes the transport object. This has been designed so that you would implement this trait on &Foo or &mut Foo instead of directly on Foo.

Associated Types

type Output

The result of a connection setup process, including protocol upgrades.

Typically the output contains at least a handle to a data stream (i.e. a connection or a substream multiplexer on top of a connection) that provides APIs for sending and receiving data through the connection.

type Error: Error

An error that occurred during connection setup.

type Listener: Stream

A stream of Outputs for inbound connections.

An item should be produced whenever a connection is received at the lowest level of the transport stack. The item must be a ListenerUpgrade future that resolves to an Output value once all protocol upgrades have been applied.

type ListenerUpgrade: Future

A pending Output for an inbound connection, obtained from the Listener stream.

After a connection has been accepted by the transport, it may need to go through asynchronous post-processing (i.e. protocol upgrade negotiations). Such post-processing should not block the Listener from producing the next connection, hence further connection setup proceeds asynchronously. Once a ListenerUpgrade future resolves it yields the Output of the connection setup process.

type Dial: Future

A pending Output for an outbound connection, obtained from dialing.

Loading content...
Required methods

fn listen_on(
self,
addr: Multiaddr
) -> Result<Self::Listener, TransportError<Self::Error>>

Listens on the given Multiaddr, producing a stream of pending, inbound connections and addresses this transport is listening on (cf. [ListenerEvent]).

fn dial(
self,
addr: Multiaddr
) -> Result<Self::Dial, TransportError<Self::Error>>

Dials the given Multiaddr, returning a future for a pending outbound connection.

If TransportError::MultiaddrNotSupported is returned, it may be desirable to try an alternative Transport, if available.

Loading content...
Provided methods

fn boxed(self) -> Boxed<Self::Output, Self::Error> where
Self: Clone + Send + Sync + 'static,
Self::Dial: Send,
Self::Dial: 'static,
Self::Listener: Send,
Self::Listener: 'static,
Self::ListenerUpgrade: Send,
Self::ListenerUpgrade: 'static,

Turns this Transport into an abstract boxed transport.

fn map<F, O>(self, map: F) -> Map<Self, F> where
F: FnOnce(Self::Output, ConnectedPoint) -> O + Clone,

Applies a function on the connections created by the transport.

fn map_err<F, TNewErr>(self, map_err: F) -> MapErr<Self, F> where
F: FnOnce(Self::Error) -> TNewErr + Clone,

Applies a function on the errors generated by the futures of the transport.

fn or_transport<T>(self, other: T) -> OrTransport<Self, T>

Builds a new transport that falls back to another transport when encountering errors on dialing or listening for connections.

The returned transport will act like self, except that if listen_on or dial return an error then other will be tried.

fn with_upgrade<U, O, E>(self, upgrade: U) -> Upgrade<Self, U> where
U: InboundUpgrade<Self::Output, Output = O, Error = E> + OutboundUpgrade<Self::Output, Output = O, Error = E>,
Self::Output: AsyncRead,
Self::Output: AsyncWrite,

Wraps this transport inside an [Upgrade].

Whenever an inbound or outbound connection is established by this transport, the upgrade is applied on the current state of the connection (which may have already gone through previous upgrades) as an [upgrade::InboundUpgrade] or [upgrade::OutboundUpgrade], respectively.

fn and_then<C, F, O>(self, upgrade: C) -> AndThen<Self, C> where
C: FnOnce(Self::Output, ConnectedPoint) -> F + Clone,
F: IntoFuture<Item = O>,

Applies a function producing an asynchronous result to every connection created by this transport.

This function can be used for ad-hoc protocol upgrades on a transport or for processing or adapting the output of an earlier upgrade before applying the next upgrade.

fn with_timeout(self, timeout: Duration) -> TransportTimeout<Self>

Adds a timeout to the connection setup (including upgrades) for all inbound and outbound connection attempts.

fn with_outbound_timeout(self, timeout: Duration) -> TransportTimeout<Self>

Adds a timeout to the connection setup (including upgrades) for all outbound connection attempts.

fn with_inbound_timeout(self, timeout: Duration) -> TransportTimeout<Self>

Adds a timeout to the connection setup (including upgrades) for all inbound connection attempts.

Loading content...