Module libp2p::core::transport::upgrade

Configuration of transport protocol upgrades.

Structs

Authenticate
An upgrade that authenticates the remote peer, typically in the context of negotiating a secure channel.

Builder
A Builder facilitates upgrading of a Transport for use with a Network.

DialUpgradeFuture
The Transport::Dial future of an Upgraded transport.

ListenerStream
The Transport::Listener stream of an Upgraded transport.

ListenerUpgradeFuture
The Transport::ListenerUpgrade future of an Upgraded transport.

Multiplex
An upgrade that negotiates a (sub)stream multiplexer on top of an authenticated transport.

Upgrade
An upgrade on an authenticated, non-multiplexed Transport.

Enums

TransportUpgradeError
Errors produced by a transport upgrade.

Version
Supported multistream-select protocol versions.