Struct libp2p::core::upgrade::MapInboundUpgrade

pub struct MapInboundUpgrade<U, F> { /* private fields */ }

๐Deprecated: Deprecated without replacement because it is not used within rust-libp2p.

Wraps around an upgrade and applies a closure to the output.

Implementations

impl<U, F> MapInboundUpgrade<U, F>

pub fn new(upgrade: U, fun: F) -> MapInboundUpgrade<U, F>