Trait libp2p::core::upgrade::OutboundUpgradeExt

pub trait OutboundUpgradeExt<C>: OutboundUpgrade<C> {
    fn map_outbound<F, T>(self, f: F) -> MapOutboundUpgrade<Self, F>
    where
        Self: Sized,
        F: FnOnce(Self::Output) -> T,
    { ... }
    fn map_outbound_err<F, T>(self, f: F) -> MapOutboundUpgradeErr<Self, F>
    where
        Self: Sized,
        F: FnOnce(Self::Error) -> T,
    { ... }
}

Extention trait for OutboundUpgrade. Automatically implemented on all types that implement OutboundUpgrade.

Provided Methods

fn map_outbound<F, T>(self, f: F) -> MapOutboundUpgrade<Self, F>where
Self: Sized,
F: FnOnce(Self::Output) -> T,

Returns a new object that wraps around Self and applies a closure to the Output.

fn map_outbound_err<F, T>(self, f: F) -> MapOutboundUpgradeErr<Self, F>where
Self: Sized,
F: FnOnce(Self::Error) -> T,

Returns a new object that wraps around Self and applies a closure to the Error.