libp2p::dns

Type Alias DnsErr

pub type DnsErr<TErr> = Error<TErr>;

๐Deprecated: Use Error instead.

Aliased Type

enum DnsErr<TErr> {
    Transport(TErr),
    ResolveError(ResolveError),
    MultiaddrNotSupported(Multiaddr),
    TooManyLookups,
}

Variants

Transport(TErr)

The underlying transport encountered an error.

ResolveError(ResolveError)

DNS resolution failed.

MultiaddrNotSupported(Multiaddr)

DNS resolution was successful, but the underlying transport refused the resolved address.

TooManyLookups

DNS resolution involved too many lookups.

DNS resolution on dialing performs up to 32 DNS lookups. If these are not sufficient to obtain a fully-resolved address, this error is returned and the DNS records for the domain(s) being dialed should be investigated.