Struct libp2p::floodsub::Floodsub

pub struct Floodsub<TSubstream> { /* fields omitted */ }

Network behaviour that automatically identifies nodes periodically, and returns information about them.

Methods

impl<TSubstream> Floodsub<TSubstream>

pub fn new(local_peer_id: PeerId) -> Floodsub<TSubstream>

Creates a Floodsub.

pub fn add_node_to_partial_view(&mut self, peer_id: PeerId)

Add a node to the list of nodes to propagate messages to.

pub fn remove_node_from_partial_view(&mut self, peer_id: &PeerId)

Remove a node from the list of nodes to propagate messages to.

impl<TSubstream> Floodsub<TSubstream>

pub fn subscribe(&mut self, topic: Topic) -> bool

Subscribes to a topic.

Returns true if the subscription worked. Returns false if we were already subscribed.

pub fn unsubscribe<impl AsRef>(
&mut self,
topic: impl AsRef
) -> bool where
impl AsRef: AsRef<TopicHash>,

Unsubscribes from a topic.

Note that this only requires a TopicHash and not a full Topic.

Returns true if we were subscribed to this topic.

pub fn publish<impl Into, impl Into>>(
&mut self,
topic: impl Into,
data: impl Into>
) where
impl Into: Into<TopicHash>,
impl Into>: Into<Vec<u8>>,

Publishes a message to the network.

Note: Doesn't do anything if we're not subscribed to the topic.

pub fn publish_many<impl Into, impl IntoIterator>, impl Into>>(
&mut self,
topic: impl IntoIterator>,
data: impl Into>
) where
impl Into: Into<TopicHash>,
impl Into>: Into<Vec<u8>>,
impl IntoIterator>: IntoIterator<Item = impl Into>,

Publishes a message with multiple topics to the network.

Note: Doesn't do anything if we're not subscribed to any of the topics.