Function libp2p::build_development_transport

pub fn build_development_transport(
    keypair: Keypair
) -> Result<Boxed<(PeerId, StreamMuxerBox)>>

Builds a Transport that supports the most commonly-used protocols that libp2p supports.

Note: This Transport is not suitable for production usage, as its implementation reserves the right to support additional protocols or remove deprecated protocols.