Function libp2p::build_tcp_ws_noise_mplex_yamux

pub fn build_tcp_ws_noise_mplex_yamux(
    keypair: Keypair
) -> Result<impl Transport<Output = (PeerId, impl StreamMuxer<OutboundSubstream = impl Send, Substream = impl Send, Error = impl Into<Error>> + Send + Sync), Error = impl Error + Send, Listener = impl Send, Dial = impl Send, ListenerUpgrade = impl Send> + Clone>

Builds an implementation of Transport that is suitable for usage with the Swarm.

The implementation supports TCP/IP, WebSockets over TCP/IP, noise as the encryption layer, and mplex or yamux as the multiplexing layer.