Function libp2p::development_transport

pub async fn development_transport(
    keypair: Keypair
) -> Result<Boxed<(PeerId, StreamMuxerBox)>>

Builds a Transport based on TCP/IP that supports the most commonly-used features of libp2p:

DNS resolution.

Noise protocol encryption.

Websockets.

Both Yamux and Mplex for substream multiplexing.

All async I/O of the transport is based on async-std.

Note: This Transport is not suitable for production usage, as its implementation reserves the right to support additional protocols or remove deprecated protocols.