Type Alias libp2p::identify::IdentifyInfo

pub type IdentifyInfo = Info;

๐Deprecated since 0.40.0: Use re-exports that omit Identify prefix, i.e. libp2p_identify::Info

Aliased Type

struct IdentifyInfo {
    pub public_key: PublicKey,
    pub protocol_version: String,
    pub agent_version: String,
    pub listen_addrs: Vec<Multiaddr>,
    pub protocols: Vec<String>,
    pub observed_addr: Multiaddr,
}

Fields
public_key: PublicKey

The public key of the local peer.

protocol_version: String

Application-specific version of the protocol family used by the peer, e.g. ipfs/1.0.0 or polkadot/1.0.0.

agent_version: String

Name and version of the peer, similar to the User-Agent header in the HTTP protocol.

listen_addrs: Vec<Multiaddr>

The addresses that the peer is listening on.

protocols: Vec<String>

The list of protocols supported by the peer, e.g. /ipfs/ping/1.0.0.

observed_addr: Multiaddr

Address observed by or for the remote.