Struct libp2p::kad::protocol::KademliaProtocolConfig

pub struct KademliaProtocolConfig { /* private fields */ }

Configuration for a Kademlia connection upgrade. When applied to a connection, turns this connection into a Stream + Sink whose items are of type KadRequestMsg and KadResponseMsg.

Implementations

impl KademliaProtocolConfig

pub fn protocol_names(&self) -> &[Cow<'static, [u8]>]

Returns the configured protocol name.

pub fn set_protocol_names(&mut self, names: Vec<Cow<'static, [u8]>>)

Modifies the protocol names used on the wire. Can be used to create incompatibilities between networks on purpose.

pub fn set_max_packet_size(&mut self, size: usize)

Modifies the maximum allowed size of a single Kademlia packet.