Type Alias libp2p::kad::KademliaEvent

pub type KademliaEvent = Event;

๐Deprecated: Import the kad module instead and refer to this type as kad::Event.

Aliased Type

enum KademliaEvent {
    InboundRequest {
        request: InboundRequest,
    },
    OutboundQueryProgressed {
        id: QueryId,
        result: QueryResult,
        stats: QueryStats,
        step: ProgressStep,
    },
    RoutingUpdated {
        peer: PeerId,
        is_new_peer: bool,
        addresses: Addresses,
        bucket_range: (Distance, Distance),
        old_peer: Option<PeerId>,
    },
    UnroutablePeer {
        peer: PeerId,
    },
    RoutablePeer {
        peer: PeerId,
        address: Multiaddr,
    },
    PendingRoutablePeer {
        peer: PeerId,
        address: Multiaddr,
    },
}

Variants

InboundRequest

Fields

request: InboundRequest

An inbound request has been received and handled.

OutboundQueryProgressed

Fields

id: QueryId

The ID of the query that finished.

result: QueryResult

The intermediate result of the query.

stats: QueryStats

Execution statistics from the query.

step: ProgressStep

Indicates which event this is, if therer are multiple responses for a single query.

An outbound query has made progress.

RoutingUpdated

Fields

peer: PeerId

The ID of the peer that was added or updated.

is_new_peer: bool

Whether this is a new peer and was thus just added to the routing table, or whether it is an existing peer whoโs addresses changed.

addresses: Addresses

The full list of known addresses of peer.

bucket_range: (Distance, Distance)

Returns the minimum inclusive and maximum inclusive distance for the bucket of the peer.

old_peer: Option<PeerId>

The ID of the peer that was evicted from the routing table to make room for the new peer, if any.

The routing table has been updated with a new peer and / or address, thereby possibly evicting another peer.

UnroutablePeer

Fields

peer: PeerId

A peer has connected for whom no listen address is known.

If the peer is to be added to the routing table, a known listen address for the peer must be provided via Behaviour::add_address.

RoutablePeer

Fields

peer: PeerId

address: Multiaddr

A connection to a peer has been established for whom a listen address is known but the peer has not been added to the routing table either because BucketInserts::Manual is configured or because the corresponding bucket is full.

If the peer is to be included in the routing table, it must must be explicitly added via Behaviour::add_address, possibly after removing another peer.

See Behaviour::kbucket for insight into the contents of the k-bucket of peer.

PendingRoutablePeer

Fields

peer: PeerId

address: Multiaddr

A connection to a peer has been established for whom a listen address is known but the peer is only pending insertion into the routing table if the least-recently disconnected peer is unresponsive, i.e. the peer may not make it into the routing table.

If the peer is to be unconditionally included in the routing table, it should be explicitly added via Behaviour::add_address after removing another peer.

See Behaviour::kbucket for insight into the contents of the k-bucket of peer.