Struct libp2p::mdns::Behaviour

pub struct Behaviour<P>where
    P: Provider,{ /* private fields */ }

A NetworkBehaviour for mDNS. Automatically discovers peers on the local network and adds them to the topology.

Implementations

impl<P> Behaviour<P>where
P: Provider,

pub fn new(config: Config) -> Result<Behaviour<P>, Error>

Builds a new Mdns behaviour.

pub fn has_node(&self, peer_id: &PeerId) -> bool

Returns true if the given PeerId is in the list of nodes discovered through mDNS.

pub fn discovered_nodes(&self) -> impl ExactSizeIterator

Returns the list of nodes that we have discovered through mDNS and that are not expired.

pub fn expire_node(&mut self, peer_id: &PeerId)

Expires a node before the ttl.