Trait libp2p::multiaddr::multihash::typenum::type_operators::Min

pub trait Min<Rhs = Self> {
    type Output;
    pub fn min(self, rhs: Rhs) -> Self::Output;
}

A type operator that returns the minimum of Self and Rhs.

Associated Types

type Output

The type of the minimum of Self and Rhs

Loading content...
Required methods

pub fn min(self, rhs: Rhs) -> Self::Output

Method returning the minimum

Loading content...