Trait libp2p::multihash::MultihashDigest

pub trait MultihashDigest {
    fn code(&self) -> Code;

    fn digest(&self, data: &[u8]) -> Multihash;
}

The MultihashDigest trait specifies an interface common for all multihash functions.

Required methods

fn code(&self) -> Code

The Mutlihash byte value.

fn digest(&self, data: &[u8]) -> Multihash

Hash some input and return the digest.

Panics

Panics if the digest length is bigger than 2^32. This only happens for identity hasing.

Loading content...