Trait libp2p::multihash::typenum::ToInt

pub trait ToInt<T> {
    fn to_int() -> T;
}

A type operator for taking a concrete integer value from a type.

It returns arbitrary integer value without explicitly specifying the type. It is useful when you pass the values to methods that accept distinct types without runtime casting.

Required methods

fn to_int() -> T

Method returning the concrete value for the type.