Module libp2p::noise::handshake

Noise protocol handshake I/O.

Structs

Handshake

A future performing a Noise handshake pattern.

Enums

IdentityExchange

The options for identity exchange in an authenticated handshake.

RemoteIdentity

The identity of the remote established during a handshake.

Functions

rt1_initiator

Creates an authenticated Noise handshake for the initiator of a single roundtrip (2 message) handshake pattern.

rt1_responder

Creates an authenticated Noise handshake for the responder of a single roundtrip (2 message) handshake pattern.

rt15_initiator

Creates an authenticated Noise handshake for the initiator of a 1.5-roundtrip (3 message) handshake pattern.

rt15_responder

Creates an authenticated Noise handshake for the responder of a 1.5-roundtrip (3 message) handshake pattern.