Struct libp2p::noise::NoiseConfig

pub struct NoiseConfig<P, C, R = ()> where
    C: Zeroize,  { /* fields omitted */ }

The protocol upgrade configuration.

Implementations

impl<H, C, R> NoiseConfig<H, C, R> where
C: Zeroize,

pub fn into_authenticated(self) -> NoiseAuthenticated<H, C, R>

Turn the NoiseConfig into an authenticated upgrade for use with a Network.

pub fn set_legacy_config(
&mut self,
cfg: LegacyConfig
) -> &mut NoiseConfig<H, C, R>

Sets the legacy configuration options to use, if any.

impl<C> NoiseConfig<IX, C, ()> where
C: Protocol<C> + Zeroize,

pub fn ix(dh_keys: AuthenticKeypair<C>) -> NoiseConfig<IX, C, ()>

Create a new NoiseConfig for the IX handshake pattern.

impl<C> NoiseConfig<XX, C, ()> where
C: Protocol<C> + Zeroize,

pub fn xx(dh_keys: AuthenticKeypair<C>) -> NoiseConfig<XX, C, ()>

Create a new NoiseConfig for the XX handshake pattern.

impl<C> NoiseConfig<IK, C, ()> where
C: Protocol<C> + Zeroize,

pub fn ik_listener(dh_keys: AuthenticKeypair<C>) -> NoiseConfig<IK, C, ()>

Create a new NoiseConfig for the IK handshake pattern (recipient side).

Since the identity of the local node is known to the remote, this configuration does not transmit a static DH public key or public identity key to the remote.

impl<C> NoiseConfig<IK, C, (PublicKey<C>, PublicKey)> where
C: Protocol<C> + Zeroize,

pub fn ik_dialer(
dh_keys: AuthenticKeypair<C>,
remote_id: PublicKey,
remote_dh: PublicKey<C>
) -> NoiseConfig<IK, C, (PublicKey<C>, PublicKey)>

Create a new NoiseConfig for the IK handshake pattern (initiator side).

In this configuration, the remote identity is known to the local node, but the local node still needs to transmit its own public identity.