Struct libp2p::ping::handler::PingHandler

pub struct PingHandler { /* fields omitted */ }

Protocol handler that handles pinging the remote at a regular period and answering ping queries.

If the remote doesn’t respond, produces an error that closes the connection.

Implementations

impl PingHandler

pub fn new(config: PingConfig) -> PingHandler

Builds a new PingHandler with the given configuration.