Type Alias libp2p::ping::PingFailure

pub type PingFailure = Failure;

๐Deprecated since 0.39.1: Use libp2p::ping::Failure instead.

Aliased Type

enum PingFailure {
    Timeout,
    Unsupported,
    Other {
        error: Box<dyn Error + Send>,
    },
}

Variants

Timeout

The ping timed out, i.e. no response was received within the configured ping timeout.

Unsupported

The peer does not support the ping protocol.

Other

The ping failed for reasons other than a timeout.

Fields

error: Box<dyn Error + Send>