Module libp2p::request_response::codec

👎Deprecated: The codec module will be made private in the future and should not be depended on.

Traits

Codec

A Codec defines the request and response types for a request-response Behaviour protocol or protocol family and how they are encoded / decoded on an I/O stream.

ProtocolName

Types serving as protocol names.

RequestResponseCodecDeprecated

A RequestResponseCodec defines the request and response types for a request-response Behaviour protocol or protocol family and how they are encoded / decoded on an I/O stream.