Struct libp2p::request_response::Behaviour

pub struct Behaviour<TCodec>where
    TCodec: Codec + Clone + Send + 'static,{ /* private fields */ }

A request/response protocol for some message codec.

Implementations

impl<TCodec> Behaviour<TCodec>where TCodec: Codec + Default + Clone + Send + 'static,

pub fn new<I>(protocols: I, cfg: Config) -> Behaviour<TCodec>where I: IntoIterator<Item = (<TCodec as Codec>::Protocol, ProtocolSupport)>,

Creates a new Behaviour for the given protocols and configuration, using Default to construct the codec.

impl<TCodec> Behaviour<TCodec>where TCodec: Codec + Clone + Send + 'static,

pub fn with_codec<I>( codec: TCodec, protocols: I, cfg: Config ) -> Behaviour<TCodec>where I: IntoIterator<Item = (<TCodec as Codec>::Protocol, ProtocolSupport)>,

Creates a new Behaviour for the given protocols, codec and configuration.

pub fn send_request( &mut self, peer: &PeerId, request: <TCodec as Codec>::Request ) -> RequestId

Initiates sending a request.

If the targeted peer is currently not connected, a dialing attempt is initiated and the request is sent as soon as a connection is established.

Note: In order for such a dialing attempt to succeed, the RequestResonse protocol must either be embedded in another NetworkBehaviour that provides peer and address discovery, or known addresses of peers must be managed via Behaviour::add_address and Behaviour::remove_address.

pub fn send_response( &mut self, ch: ResponseChannel<<TCodec as Codec>::Response>, rs: <TCodec as Codec>::Response ) -> Result<(), <TCodec as Codec>::Response>

Initiates sending a response to an inbound request.

If the ResponseChannel is already closed due to a timeout or the connection being closed, the response is returned as an Err for further handling. Once the response has been successfully sent on the corresponding connection, Event::ResponseSent is emitted. In all other cases Event::InboundFailure will be or has been emitted.

The provided ResponseChannel is obtained from an inbound Message::Request.

pub fn add_address(&mut self, peer: &PeerId, address: Multiaddr)

Adds a known address for a peer that can be used for dialing attempts by the Swarm, i.e. is returned by NetworkBehaviour::handle_pending_outbound_connection.

Addresses added in this way are only removed by remove_address.

pub fn remove_address(&mut self, peer: &PeerId, address: &Multiaddr)

Removes an address of a peer previously added via add_address.

pub fn is_connected(&self, peer: &PeerId) -> bool

Checks whether a peer is currently connected.

pub fn is_pending_outbound(&self, peer: &PeerId, request_id: &RequestId) -> bool

Checks whether an outbound request to the peer with the provided PeerId initiated by Behaviour::send_request is still pending, i.e. waiting for a response.

pub fn is_pending_inbound(&self, peer: &PeerId, request_id: &RequestId) -> bool

Checks whether an inbound request from the peer with the provided PeerId is still pending, i.e. waiting for a response by the local node through Behaviour::send_response.