Struct libp2p::request_response::Throttled

pub struct Throttled<C> where
    C: RequestResponseCodec + Send,
    <C as RequestResponseCodec>::Protocol: Sync,  { /* fields omitted */ }

A wrapper around RequestResponse which adds request limits per peer.

Implementations

impl<C> Throttled<C> where
C: RequestResponseCodec + Send + Clone,
<C as RequestResponseCodec>::Protocol: Sync,

pub fn new<I>(c: C, protos: I, cfg: RequestResponseConfig) -> Throttled<C> where
C: Send,
I: IntoIterator<Item = (<C as RequestResponseCodec>::Protocol, ProtocolSupport)>,
<C as RequestResponseCodec>::Protocol: Sync,

Create a new throttled request-response behaviour.

pub fn from(behaviour: RequestResponse<Codec<C>>) -> Throttled<C>

Wrap an existing RequestResponse behaviour and apply send/recv limits.

pub fn set_receive_limit(&mut self, limit: NonZeroU16)

Set the global default receive limit per peer.

pub fn override_receive_limit(&mut self, p: &PeerId, limit: NonZeroU16)

Override the receive limit of a single peer.

pub fn remove_override(&mut self, p: &PeerId)

Remove any limit overrides for the given peer.

pub fn can_send(&mut self, p: &PeerId) -> bool

Has the limit of outbound requests been reached for the given peer?

pub fn send_request(
&mut self,
p: &PeerId,
req: <C as RequestResponseCodec>::Request
) -> Result<RequestId, <C as RequestResponseCodec>::Request>

Send a request to a peer.

If the limit of outbound requests has been reached, the request is returned. Sending more outbound requests should only be attempted once Event::ResumeSending has been received from NetworkBehaviour::poll.

pub fn send_response(
&mut self,
ch: ResponseChannel<Message<<C as RequestResponseCodec>::Response>>,
res: <C as RequestResponseCodec>::Response
)

Answer an inbound request with a response.

See RequestResponse::send_response for details.

pub fn add_address(&mut self, p: &PeerId, a: Multiaddr)

Add a known peer address.

See RequestResponse::add_address for details.

pub fn remove_address(&mut self, p: &PeerId, a: &Multiaddr)

Remove a previously added peer address.

See RequestResponse::remove_address for details.

pub fn is_connected(&self, p: &PeerId) -> bool

Are we connected to the given peer?

See RequestResponse::is_connected for details.

pub fn is_pending_outbound(&self, p: &RequestId) -> bool

Are we waiting for a response to the given request?

See RequestResponse::is_pending_outbound for details.