Type Alias libp2p::request_response::HandlerEvent

pub type HandlerEvent<TCodec> = Event<TCodec>;

👎Deprecated since 0.24.0: Use re-exports that omit RequestResponse prefix, i.e. libp2p::request_response::handler::Event

Aliased Type

enum HandlerEvent<TCodec> {
    Request {
        request_id: RequestId,
        request: <TCodec as Codec>::Request,
        sender: Sender<<TCodec as Codec>::Response>,
    },
    Response {
        request_id: RequestId,
        response: <TCodec as Codec>::Response,
    },
    ResponseSent(RequestId),
    ResponseOmission(RequestId),
    OutboundTimeout(RequestId),
    OutboundUnsupportedProtocols(RequestId),
    InboundTimeout(RequestId),
    InboundUnsupportedProtocols(RequestId),
}

Variants

Request

Fields

request_id: RequestId

request: <TCodec as Codec>::Request

sender: Sender<<TCodec as Codec>::Response>

A request has been received.

Response

Fields

request_id: RequestId

response: <TCodec as Codec>::Response

A response has been received.

ResponseSent(RequestId)

A response to an inbound request has been sent.

ResponseOmission(RequestId)

A response to an inbound request was omitted as a result of dropping the response sender of an inbound Request.

OutboundTimeout(RequestId)

An outbound request timed out while sending the request or waiting for the response.

OutboundUnsupportedProtocols(RequestId)

An outbound request failed to negotiate a mutually supported protocol.

InboundTimeout(RequestId)

An inbound request timed out while waiting for the request or sending the response.

InboundUnsupportedProtocols(RequestId)

An inbound request failed to negotiate a mutually supported protocol.