Struct libp2p::swarm::behaviour::toggle::Toggle

pub struct Toggle<TBehaviour> { /* private fields */ }

Implementation of NetworkBehaviour that can be either in the disabled or enabled state.

The state can only be chosen at initialization.

Implementations

impl<TBehaviour> Toggle<TBehaviour>

pub fn is_enabled(&self) -> bool

Returns true if Toggle is enabled and false if itโs disabled.

pub fn as_ref(&self) -> Option<&TBehaviour>

Returns a reference to the inner NetworkBehaviour.

pub fn as_mut(&mut self) -> Option<&mut TBehaviour>

Returns a mutable reference to the inner NetworkBehaviour.