Trait libp2p::swarm::behaviour::PollParameters

pub trait PollParameters {
    type SupportedProtocolsIter: ExactSizeIterator<Item = Vec<u8>>;
    type ListenedAddressesIter: ExactSizeIterator<Item = Multiaddr>;
    type ExternalAddressesIter: ExactSizeIterator<Item = AddressRecord>;

    // Required methods
    fn supported_protocols(&self) -> Self::SupportedProtocolsIter;
    fn listened_addresses(&self) -> Self::ListenedAddressesIter;
    fn external_addresses(&self) -> Self::ExternalAddressesIter;
    fn local_peer_id(&self) -> &PeerId;
}

Parameters passed to poll(), that the NetworkBehaviour has access to.

Required Associated Types

type SupportedProtocolsIter: ExactSizeIterator<Item = Vec<u8>>

Iterator returned by supported_protocols.

type ListenedAddressesIter: ExactSizeIterator<Item = Multiaddr>

Iterator returned by listened_addresses.

type ExternalAddressesIter: ExactSizeIterator<Item = AddressRecord>

Iterator returned by external_addresses.

Required Methods

fn supported_protocols(&self) -> Self::SupportedProtocolsIter

Returns the list of protocol the behaviour supports when a remote negotiates a protocol on an inbound substream.

The iterator’s elements are the ASCII names as reported on the wire.

Note that the list is computed once at initialization and never refreshed.

fn listened_addresses(&self) -> Self::ListenedAddressesIter

👎Deprecated since 0.42.0: Use libp2p_swarm::ListenAddresses instead.

Returns the list of the addresses we’re listening on.

fn external_addresses(&self) -> Self::ExternalAddressesIter

👎Deprecated since 0.42.0: Use libp2p_swarm::ExternalAddresses instead.

Returns the list of the addresses nodes can use to reach us.

fn local_peer_id(&self) -> &PeerId

👎Deprecated since 0.42.0: Pass the node’s PeerId into the behaviour instead.

Returns the peer id of the local node.