Enum libp2p::swarm::ToSwarm

pub enum ToSwarm<TOutEvent, TInEvent> {
    GenerateEvent(TOutEvent),
    Dial {
        opts: DialOpts,
    },
    ListenOn {
        opts: ListenOpts,
    },
    RemoveListener {
        id: ListenerId,
    },
    NotifyHandler {
        peer_id: PeerId,
        handler: NotifyHandler,
        event: TInEvent,
    },
    NewExternalAddrCandidate(Multiaddr),
    ExternalAddrConfirmed(Multiaddr),
    ExternalAddrExpired(Multiaddr),
    CloseConnection {
        peer_id: PeerId,
        connection: CloseConnection,
    },
}

A command issued from a NetworkBehaviour for the Swarm.

Variants

GenerateEvent(TOutEvent)

Instructs the Swarm to return an event when it is being polled.

Dial

Fields

opts: DialOpts

Instructs the swarm to start a dial.

On success, NetworkBehaviour::on_swarm_event with ConnectionEstablished is invoked. On failure, NetworkBehaviour::on_swarm_event with DialFailure is invoked.

DialOpts provides access to the ConnectionId via DialOpts::connection_id. This ConnectionId will be used throughout the connection’s lifecycle to associate events with it. This allows a NetworkBehaviour to identify a connection that resulted out of its own dial request.

ListenOn

Fields

opts: ListenOpts

Instructs the Swarm to listen on the provided address.

RemoveListener

Fields

id: ListenerId

Instructs the Swarm to remove the listener.

NotifyHandler

Fields

peer_id: PeerId

The peer for whom a ConnectionHandler should be notified.

handler: NotifyHandler

The options w.r.t. which connection handler to notify of the event.

event: TInEvent

The event to send.

Instructs the Swarm to send an event to the handler dedicated to a connection with a peer.

If the Swarm is connected to the peer, the message is delivered to the ConnectionHandler instance identified by the peer ID and connection ID.

If the specified connection no longer exists, the event is silently dropped.

Typically the connection ID given is the same as the one passed to NetworkBehaviour::on_connection_handler_event, i.e. whenever the behaviour wishes to respond to a request on the same connection (and possibly the same substream, as per the implementation of ConnectionHandler).

Note that even if the peer is currently connected, connections can get closed at any time and thus the event may not reach a handler.

NewExternalAddrCandidate(Multiaddr)

Reports a new candidate for an external address to the Swarm.

This address will be shared with all NetworkBehaviours via FromSwarm::NewExternalAddrCandidate.

This address could come from a variety of sources:

A protocol such as identify obtained it from a remote.

The user provided it based on configuration.

We made an educated guess based on one of our listen addresses.

We established a new relay connection.

ExternalAddrConfirmed(Multiaddr)

Indicates to the Swarm that the provided address is confirmed to be externally reachable.

This is intended to be issued in response to a FromSwarm::NewExternalAddrCandidate if we are indeed externally reachable on this address. This address will be shared with all NetworkBehaviours via FromSwarm::ExternalAddrConfirmed.

ExternalAddrExpired(Multiaddr)

Indicates to the Swarm that we are no longer externally reachable under the provided address.

This expires an address that was earlier confirmed via ToSwarm::ExternalAddrConfirmed. This address will be shared with all NetworkBehaviours via FromSwarm::ExternalAddrExpired.

CloseConnection

Fields

peer_id: PeerId

The peer to disconnect.

connection: CloseConnection

Whether to close a specific or all connections to the given peer.

Instructs the Swarm to initiate a graceful close of one or all connections with the given peer.

Note: Closing a connection via ToSwarm::CloseConnection does not inform the corresponding ConnectionHandler. Closing a connection via a ConnectionHandler can be done either in a collaborative manner across ConnectionHandlers with ConnectionHandler::connection_keep_alive or directly with ConnectionHandlerEvent::Close.

Implementations

impl<TOutEvent, TInEventOld> ToSwarm<TOutEvent, TInEventOld>

pub fn map_in<TInEventNew>( self, f: impl FnOnce(TInEventOld) -> TInEventNew ) -> ToSwarm<TOutEvent, TInEventNew>

Map the handler event.

impl<TOutEvent, THandlerIn> ToSwarm<TOutEvent, THandlerIn>

pub fn map_out<E>( self, f: impl FnOnce(TOutEvent) -> E ) -> ToSwarm<E, THandlerIn>

Map the event the swarm will return.