Trait libp2p::swarm::handler::IntoConnectionHandler

pub trait IntoConnectionHandler: 'static + Send {
    type Handler: ConnectionHandler;

    fn into_handler(
        self,
        remote_peer_id: &PeerId,
        connected_point: &ConnectedPoint
    ) -> Self::Handler;
    fn inbound_protocol(
        &self
    ) -> <Self::Handler as ConnectionHandler>::InboundProtocol;

    fn select<TProto2>(
        self,
        other: TProto2
    ) -> IntoConnectionHandlerSelect<Self, TProto2>
    where
        Self: Sized,
    { ... }
}

Prototype for a ConnectionHandler.

Required Associated Types

type Handler: ConnectionHandler

The protocols handler.

Required Methods

fn into_handler(
self,
remote_peer_id: &PeerId,
connected_point: &ConnectedPoint
) -> Self::Handler

Builds the protocols handler.

The PeerId is the id of the node the handler is going to handle.

fn inbound_protocol(
&self
) -> <Self::Handler as ConnectionHandler>::InboundProtocol

Return the handler’s inbound protocol.

Provided Methods

fn select<TProto2>(
self,
other: TProto2
) -> IntoConnectionHandlerSelect<Self, TProto2>where
Self: Sized,

Builds an implementation of IntoConnectionHandler that handles both this protocol and the other one together.