libp2p::swarm

Struct Swarm

pub struct Swarm<TBehaviour>
where
    TBehaviour: NetworkBehaviour,
{ /* private fields */ }

Contains the state of the network, plus the way it should behave.

Note: Needs to be polled via <Swarm as Stream> in order to make progress.

Implementations

impl<TBehaviour> Swarm<TBehaviour>
where TBehaviour: NetworkBehaviour,

pub fn new( transport: Boxed<(PeerId, StreamMuxerBox)>, behaviour: TBehaviour, local_peer_id: PeerId, config: Config, ) -> Swarm<TBehaviour>

Creates a new Swarm from the given Transport, NetworkBehaviour, PeerId and Config.

pub fn network_info(&self) -> NetworkInfo

Returns information about the connections underlying the Swarm.

pub fn listen_on( &mut self, addr: Multiaddr, ) -> Result<ListenerId, TransportError<Error>>

Starts listening on the given address. Returns an error if the address is not supported.

Listeners report their new listening addresses as SwarmEvent::NewListenAddr. Depending on the underlying transport, one listener may have multiple listening addresses.

pub fn remove_listener(&mut self, listener_id: ListenerId) -> bool

Remove some listener.

Returns true if there was a listener with this ID, false otherwise.

pub fn dial(&mut self, opts: impl Into<DialOpts>) -> Result<(), DialError>

Dial a known or unknown peer.

See also DialOpts.

let mut swarm = SwarmBuilder::with_tokio_executor(
    DummyTransport::new().boxed(),
    dummy::Behaviour,
    PeerId::random(),
).build();

// Dial a known peer.
swarm.dial(PeerId::random());

// Dial an unknown peer.
swarm.dial("/ip6/::1/tcp/12345".parse::<Multiaddr>().unwrap());

pub fn listeners(&self) -> impl Iterator<Item = &Multiaddr>

Returns an iterator that produces the list of addresses we’re listening on.

pub fn local_peer_id(&self) -> &PeerId

Returns the peer ID of the swarm passed as parameter.

pub fn external_addresses(&self) -> impl Iterator<Item = &Multiaddr>

List all confirmed external address for the local node.

pub fn add_external_address(&mut self, a: Multiaddr)

Add a confirmed external address for the local node.

This function should only be called with addresses that are guaranteed to be reachable. The address is broadcast to all NetworkBehaviours via FromSwarm::ExternalAddrConfirmed.

pub fn remove_external_address(&mut self, addr: &Multiaddr)

Remove an external address for the local node.

The address is broadcast to all NetworkBehaviours via FromSwarm::ExternalAddrExpired.

pub fn disconnect_peer_id(&mut self, peer_id: PeerId) -> Result<(), ()>

Disconnects a peer by its peer ID, closing all connections to said peer.

Returns Ok(()) if there was one or more established connections to the peer.

Note: Closing a connection via Swarm::disconnect_peer_id does not inform the corresponding ConnectionHandler. Closing a connection via a ConnectionHandler can be done either in a collaborative manner across ConnectionHandlers with ConnectionHandler::connection_keep_alive or directly with ConnectionHandlerEvent::Close.

pub fn close_connection(&mut self, connection_id: ConnectionId) -> bool

Attempt to gracefully close a connection.

Closing a connection is asynchronous but this function will return immediately. A SwarmEvent::ConnectionClosed event will be emitted once the connection is actually closed.

Returns

true if the connection was established and is now being closed.

false if the connection was not found or is no longer established.

pub fn is_connected(&self, peer_id: &PeerId) -> bool

Checks whether there is an established connection to a peer.

pub fn connected_peers(&self) -> impl Iterator<Item = &PeerId>

Returns the currently connected peers.

pub fn behaviour(&self) -> &TBehaviour

Returns a reference to the provided NetworkBehaviour.

pub fn behaviour_mut(&mut self) -> &mut TBehaviour

Returns a mutable reference to the provided NetworkBehaviour.