Trait libp2p::swarm::PollParameters

pub trait PollParameters {
    type SupportedProtocolsIter: ExactSizeIterator;
    type ListenedAddressesIter: ExactSizeIterator;
    type ExternalAddressesIter: ExactSizeIterator;
    pub fn supported_protocols(&self) -> Self::SupportedProtocolsIter;

    pub fn listened_addresses(&self) -> Self::ListenedAddressesIter;

    pub fn external_addresses(&self) -> Self::ExternalAddressesIter;

    pub fn local_peer_id(&self) -> &PeerId;
}

Parameters passed to poll(), that the NetworkBehaviour has access to.

Associated Types

type SupportedProtocolsIter: ExactSizeIterator

Iterator returned by supported_protocols.

type ListenedAddressesIter: ExactSizeIterator

Iterator returned by listened_addresses.

type ExternalAddressesIter: ExactSizeIterator

Iterator returned by external_addresses.

Loading content...
Required methods

pub fn supported_protocols(&self) -> Self::SupportedProtocolsIter

Returns the list of protocol the behaviour supports when a remote negotiates a protocol on an inbound substream.

The iterator’s elements are the ASCII names as reported on the wire.

Note that the list is computed once at initialization and never refreshed.

pub fn listened_addresses(&self) -> Self::ListenedAddressesIter

Returns the list of the addresses we’re listening on.

pub fn external_addresses(&self) -> Self::ExternalAddressesIter

Returns the list of the addresses nodes can use to reach us.

pub fn local_peer_id(&self) -> &PeerId

Returns the peer id of the local node.

Loading content...