Trait libp2p::swarm::ProtocolsHandler

pub trait ProtocolsHandler: 'static + Send {
    type InEvent: 'static + Send;
    type OutEvent: 'static + Send;
    type Error: 'static + Error + Send;
    type InboundProtocol: InboundUpgradeSend;
    type OutboundProtocol: OutboundUpgradeSend;
    type InboundOpenInfo: 'static + Send;
    type OutboundOpenInfo: 'static + Send;

    pub fn listen_protocol(
        &self
    ) -> SubstreamProtocol<Self::InboundProtocol, Self::InboundOpenInfo>;

    pub fn inject_fully_negotiated_inbound(
        &mut self, 
        protocol: <Self::InboundProtocol as InboundUpgradeSend>::Output, 
        info: Self::InboundOpenInfo
    );

    pub fn inject_fully_negotiated_outbound(
        &mut self, 
        protocol: <Self::OutboundProtocol as OutboundUpgradeSend>::Output, 
        info: Self::OutboundOpenInfo
    );

    pub fn inject_event(&mut self, event: Self::InEvent);

    pub fn inject_dial_upgrade_error(
        &mut self, 
        info: Self::OutboundOpenInfo, 
        error: ProtocolsHandlerUpgrErr<<Self::OutboundProtocol as OutboundUpgradeSend>::Error>
    );

    pub fn connection_keep_alive(&self) -> KeepAlive;

    pub fn poll(
        &mut self, 
        cx: &mut Context<'_>
    ) -> Poll<ProtocolsHandlerEvent<Self::OutboundProtocol, Self::OutboundOpenInfo, Self::OutEvent, Self::Error>>;

    pub fn inject_address_change(&mut self, _new_address: &Multiaddr) { ... }

    pub fn inject_listen_upgrade_error(
        &mut self, 
        Self::InboundOpenInfo, 
        ProtocolsHandlerUpgrErr<<Self::InboundProtocol as InboundUpgradeSend>::Error>
    ) { ... }

    pub fn map_in_event<TNewIn, TMap>(
        self, 
        map: TMap
    ) -> MapInEvent<Self, TNewIn, TMap>
    where
        TMap: Fn(&TNewIn) -> Option<&Self::InEvent>,
    { ... }

    pub fn map_out_event<TMap, TNewOut>(
        self, 
        map: TMap
    ) -> MapOutEvent<Self, TMap>
    where
        TMap: FnMut(Self::OutEvent) -> TNewOut,
    { ... }

    pub fn select<TProto2>(
        self, 
        other: TProto2
    ) -> ProtocolsHandlerSelect<Self, TProto2> { ... }

    pub fn into_node_handler_builder(self) -> NodeHandlerWrapperBuilder<Self> { ... }

}

A handler for a set of protocols used on a connection with a remote.

This trait should be implemented for a type that maintains the state for the execution of a specific protocol with a remote.

Handling a protocol

Communication with a remote over a set of protocols is initiated in one of two ways:

Dialing by initiating a new outbound substream. In order to do so, ProtocolsHandler::poll() must return an ProtocolsHandlerEvent::OutboundSubstreamRequest, providing an instance of libp2p_core::upgrade::OutboundUpgrade that is used to negotiate the protocol(s). Upon success, ProtocolsHandler::inject_fully_negotiated_outbound is called with the final output of the upgrade.

Listening by accepting a new inbound substream. When a new inbound substream is created on a connection, ProtocolsHandler::listen_protocol is called to obtain an instance of libp2p_core::upgrade::InboundUpgrade that is used to negotiate the protocol(s). Upon success, ProtocolsHandler::inject_fully_negotiated_inbound is called with the final output of the upgrade.

Connection Keep-Alive

A ProtocolsHandler can influence the lifetime of the underlying connection through ProtocolsHandler::connection_keep_alive. That is, the protocol implemented by the handler can include conditions for terminating the connection. The lifetime of successfully negotiated substreams is fully controlled by the handler.

Implementors of this trait should keep in mind that the connection can be closed at any time. When a connection is closed gracefully, the substreams used by the handler may still continue reading data until the remote closes its side of the connection.

Associated Types

type InEvent: 'static + Send

Custom event that can be received from the outside.

type OutEvent: 'static + Send

Custom event that can be produced by the handler and that will be returned to the outside.

type Error: 'static + Error + Send

The type of errors returned by ProtocolsHandler::poll.

type InboundProtocol: InboundUpgradeSend

The inbound upgrade for the protocol(s) used by the handler.

type OutboundProtocol: OutboundUpgradeSend

The outbound upgrade for the protocol(s) used by the handler.

type InboundOpenInfo: 'static + Send

The type of additional information returned from listen_protocol.

type OutboundOpenInfo: 'static + Send

The type of additional information passed to an OutboundSubstreamRequest.

Loading content...
Required methods

pub fn listen_protocol(
&self
) -> SubstreamProtocol<Self::InboundProtocol, Self::InboundOpenInfo>

The InboundUpgrade to apply on inbound substreams to negotiate the desired protocols.

Note: The returned InboundUpgrade should always accept all the generally supported protocols, even if in a specific context a particular one is not supported, (eg. when only allowing one substream at a time for a protocol). This allows a remote to put the list of supported protocols in a cache.

pub fn inject_fully_negotiated_inbound(
&mut self,
protocol: <Self::InboundProtocol as InboundUpgradeSend>::Output,
info: Self::InboundOpenInfo
)

Injects the output of a successful upgrade on a new inbound substream.

pub fn inject_fully_negotiated_outbound(
&mut self,
protocol: <Self::OutboundProtocol as OutboundUpgradeSend>::Output,
info: Self::OutboundOpenInfo
)

Injects the output of a successful upgrade on a new outbound substream.

The second argument is the information that was previously passed to ProtocolsHandlerEvent::OutboundSubstreamRequest.

pub fn inject_event(&mut self, event: Self::InEvent)

Injects an event coming from the outside in the handler.

pub fn inject_dial_upgrade_error(
&mut self,
info: Self::OutboundOpenInfo,
error: ProtocolsHandlerUpgrErr<<Self::OutboundProtocol as OutboundUpgradeSend>::Error>
)

Indicates to the handler that upgrading an outbound substream to the given protocol has failed.

pub fn connection_keep_alive(&self) -> KeepAlive

Returns until when the connection should be kept alive.

This method is called by the Swarm after each invocation of ProtocolsHandler::poll to determine if the connection and the associated ProtocolsHandlers should be kept alive as far as this handler is concerned and if so, for how long.

Returning KeepAlive::No indicates that the connection should be closed and this handler destroyed immediately.

Returning KeepAlive::Until indicates that the connection may be closed and this handler destroyed after the specified Instant.

Returning KeepAlive::Yes indicates that the connection should be kept alive until the next call to this method.

Note: The connection is always closed and the handler destroyed when ProtocolsHandler::poll returns an error. Furthermore, the connection may be closed for reasons outside of the control of the handler.

pub fn poll(
&mut self,
cx: &mut Context<'_>
) -> Poll<ProtocolsHandlerEvent<Self::OutboundProtocol, Self::OutboundOpenInfo, Self::OutEvent, Self::Error>>

Should behave like Stream::poll().

Loading content...
Provided methods

pub fn inject_address_change(&mut self, _new_address: &Multiaddr)

Notifies the handler of a change in the address of the remote.

pub fn inject_listen_upgrade_error(
&mut self,
Self::InboundOpenInfo,
ProtocolsHandlerUpgrErr<<Self::InboundProtocol as InboundUpgradeSend>::Error>
)

Indicates to the handler that upgrading an inbound substream to the given protocol has failed.

pub fn map_in_event<TNewIn, TMap>(
self,
map: TMap
) -> MapInEvent<Self, TNewIn, TMap> where
TMap: Fn(&TNewIn) -> Option<&Self::InEvent>,

Adds a closure that turns the input event into something else.

pub fn map_out_event<TMap, TNewOut>(self, map: TMap) -> MapOutEvent<Self, TMap> where
TMap: FnMut(Self::OutEvent) -> TNewOut,

Adds a closure that turns the output event into something else.

pub fn select<TProto2>(
self,
other: TProto2
) -> ProtocolsHandlerSelect<Self, TProto2>

Creates a new ProtocolsHandler that selects either this handler or other by delegating methods calls appropriately.

Note: The largest KeepAlive returned by the two handlers takes precedence, i.e. is returned from ProtocolsHandler::connection_keep_alive by the returned handler.

pub fn into_node_handler_builder(self) -> NodeHandlerWrapperBuilder<Self>

Creates a builder that allows creating a NodeHandler that handles this protocol exclusively.

Note: This method should not be redefined in a custom ProtocolsHandler.

Loading content...