Trait libp2p::InboundUpgradeExt

pub trait InboundUpgradeExt<C>: InboundUpgrade<C> {
    pub fn map_inbound<F, T>(self, f: F) -> MapInboundUpgrade<Self, F>
    where
        F: FnOnce(Self::Output) -> T,
    { ... }

    pub fn map_inbound_err<F, T>(self, f: F) -> MapInboundUpgradeErr<Self, F>
    where
        F: FnOnce(Self::Error) -> T,
    { ... }
}

Extension trait for InboundUpgrade. Automatically implemented on all types that implement InboundUpgrade.

Provided methods

pub fn map_inbound<F, T>(self, f: F) -> MapInboundUpgrade<Self, F> where
F: FnOnce(Self::Output) -> T,

Returns a new object that wraps around Self and applies a closure to the Output.

pub fn map_inbound_err<F, T>(self, f: F) -> MapInboundUpgradeErr<Self, F> where
F: FnOnce(Self::Error) -> T,

Returns a new object that wraps around Self and applies a closure to the Error.

Loading content...