libp2p::websocket

Module framed

Structs

Connection

The websocket connection.

WsConfig

A Websocket transport whose output type is a Stream and Sink of frame payloads which does not implement AsyncRead or AsyncWrite. See crate::WsConfig if you require the latter.

Enums

Data

Application data received over the websocket connection

Incoming

Data or control information received over the websocket connection.

OutgoingData

Data sent over the websocket connection.