Struct libp2p::websocket::framed::WsConfig

pub struct WsConfig<T> { /* fields omitted */ }

A Websocket transport whose output type is a Stream and Sink of frame payloads which does not implement AsyncRead or AsyncWrite. See crate::WsConfig if you require the latter.

Implementations

impl<T> WsConfig<T>

pub fn new(transport: T) -> WsConfig<T>

Create a new websocket transport based on another transport.

pub fn max_redirects(&self) -> u8

Return the configured maximum number of redirects.

pub fn set_max_redirects(&mut self, max: u8) -> &mut WsConfig<T>

Set max. number of redirects to follow.

pub fn max_data_size(&self) -> usize

Get the max. frame data size we support.

pub fn set_max_data_size(&mut self, size: usize) -> &mut WsConfig<T>

Set the max. frame data size we support.

pub fn set_tls_config(&mut self, c: Config) -> &mut WsConfig<T>

Set the TLS configuration if TLS support is desired.

pub fn use_deflate(&mut self, flag: bool) -> &mut WsConfig<T>

Should the deflate extension (RFC 7692) be used if supported?