Struct libp2p::websocket::WsConfig

pub struct WsConfig<T>where
    T: Transport,
    <T as Transport>::Output: 'static + AsyncRead + AsyncWrite + Send + Unpin,{ /* private fields */ }

A Websocket transport.

Implementations

impl<T> WsConfig<T>where
T: Transport + 'static + Send + Unpin,
<T as Transport>::Error: 'static + Send,
<T as Transport>::Dial: 'static + Send,
<T as Transport>::ListenerUpgrade: 'static + Send,
<T as Transport>::Output: 'static + AsyncRead + AsyncWrite + Send + Unpin,

pub fn new(transport: T) -> WsConfig<T>

Create a new websocket transport based on the given transport.

*Note: The given transport must be based on TCP/IP and should usually incorporate DNS resolution, though the latter is not strictly necessary if one wishes to only use the Ws protocol with known IP addresses and ports. See libp2p-tcp and libp2p-dns for constructing the inner transport.

pub fn max_redirects(&self) -> u8

Return the configured maximum number of redirects.

pub fn set_max_redirects(&mut self, max: u8) -> &mut WsConfig<T>

Set max. number of redirects to follow.

pub fn max_data_size(&self) -> usize

Get the max. frame data size we support.

pub fn set_max_data_size(&mut self, size: usize) -> &mut WsConfig<T>

Set the max. frame data size we support.

pub fn set_tls_config(&mut self, c: Config) -> &mut WsConfig<T>

Set the TLS configuration if TLS support is desired.

pub fn use_deflate(&mut self, flag: bool) -> &mut WsConfig<T>

Should the deflate extension (RFC 7692) be used if supported?