libp2p

Crate yamux

Implementation of the Yamux multiplexing protocol for libp2p.

Structs

Config

The yamux configuration.

Error

The Yamux StreamMuxer error type.

Muxer

A Yamux connection.

Stream

A stream produced by the yamux multiplexer.

WindowUpdateMode

The window update mode determines when window updates are sent to the remote, giving it new credit to send more data.