Crate libp2p

libp2p is a modular peer-to-peer networking framework.

To learn more about the general libp2p multi-language framework visit https://libp2p.io.

To get started with this libp2p implementation in Rust, please take a look at the tutorials. Further examples can be found in the examples directory.

Re-exports

pub use bytes;

pub use futures;

pub use libp2p_quic as quic;

pub use libp2p_identity as identity;

Modules

allow_block_list

A libp2p module for managing allow and blocks lists to peers.

bandwidth

connection_limits

core

Transports, upgrades, multiplexing and node handling of libp2p.

dns

DNS name resolution

gossipsub

Implementation of the Gossipsub protocol.

identify

Implementation of the Identify protocol.

kad

Implementation of the libp2p-specific Kademlia protocol.

multiaddr

Implementation of multiaddr in Rust.

multihash

Bare-minimum multihash data structure.

noise

Noise protocol framework support for libp2p.

ping

This module implements the /ipfs/ping/1.0.0 protocol.

request_response

Generic request/response protocols.

swarm

High-level network manager.

tcp

Implementation of the libp2p libp2p_core::Transport trait for TCP/IP.

tls

TLS configuration based on libp2p TLS specs.

tutorials

Rust-libp2p Tutorials to get started with.

yamux

Implementation of the Yamux multiplexing protocol for libp2p.

Macros

build_multiaddr

Easy way for a user to create a Multiaddr.

Structs

Multiaddr

Representation of a Multiaddr.

PeerId

Identifier of a peer of the network.

Stream

StreamProtocol

Identifies a protocol for a stream.

Swarm

Contains the state of the network, plus the way it should behave.

SwarmBuilder

Build a Swarm by combining an identity, a set of Transports and a NetworkBehaviour.

Enums

TransportError

An error during dialing or listening on a Transport.

Traits

InboundUpgrade

Possible upgrade on an inbound connection or substream.

OutboundUpgrade

Possible upgrade on an outbound connection or substream.

Transport

A transport provides connection-oriented communication between two peers through ordered streams of data (i.e. connections).

TransportExt

Trait automatically implemented on all objects that implement Transport. Provides some additional utilities.