Crate libp2p::core

Transports, upgrades, multiplexing and node handling of libp2p.

The main concepts of libp2p-core are:

A PeerId is a unique global identifier for a node on the network. Each node must have a different PeerId. Normally, a PeerId is the hash of the public key used to negotiate encryption on the communication channel, thereby guaranteeing that they cannot be spoofed.

The Transport trait defines how to reach a remote node or listen for incoming remote connections. See the transport module.

The StreamMuxer trait is implemented on structs that hold a connection to a remote and can subdivide this connection into multiple substreams. See the muxing module.

The UpgradeInfo, InboundUpgrade and OutboundUpgrade traits define how to upgrade each individual substream to use a protocol. See the upgrade module.

Modules

connection

either

identity
A node’s network identity keys.

multiaddr

muxing
Muxing is the process of splitting a connection into multiple substreams.

network

transport
Connection-oriented communication channels.

upgrade
Contains everything related to upgrading a connection or a substream to use a protocol.

Structs

Connected
Information about a successfully established connection.

Multiaddr
Representation of a Multiaddr.

Network
Implementation of Stream that handles the nodes.

PeerId
Identifier of a peer of the network.

Enums

ConnectedPoint
The endpoint roles associated with a peer-to-peer connection.

Endpoint
The endpoint roles associated with a peer-to-peer communication channel.

PublicKey
The public key of a node’s identity keypair.

UpgradeError
Error that can happen when upgrading a connection or substream to use a protocol.

Traits

Executor
Implemented on objects that can run a Future in the background.

InboundUpgrade
Possible upgrade on an inbound connection or substream.

OutboundUpgrade
Possible upgrade on an outbound connection or substream.

ProtocolName
Types serving as protocol names.

StreamMuxer
Implemented on objects that can open and manage substreams.

Transport
A transport provides connection-oriented communication between two peers through ordered streams of data (i.e. connections).

UpgradeInfo
Common trait for upgrades that can be applied on inbound substreams, outbound substreams, or both.

Functions

address_translation
Perform IP address translation.

Type Definitions

Negotiated