Trait libp2p::core::connection::IntoConnectionHandler

pub trait IntoConnectionHandler {
    type Handler: ConnectionHandler;
    fn into_handler(self, connected: &Connected) -> Self::Handler;
}

Prototype for a ConnectionHandler.

Associated Types

type Handler: ConnectionHandler

The node handler.

Required methods

fn into_handler(self, connected: &Connected) -> Self::Handler

Builds the node handler.

The implementation is given a Connected value that holds information about the newly established connection for which a handler should be created.