Module libp2p::core::nodes

Low-level networking primitives.

Contains structs that are aiming at providing very precise control over what happens over the network.

The more complete and highest-level struct is the RawSwarm. The RawSwarm directly or indirectly uses all the other structs of this module.

Modules

collection

handled_node

handled_node_tasks

listeners
Manage listening on multiple multiaddresses at once.

node

raw_swarm

Structs

RawSwarm
Implementation of Stream that handles the nodes.

Enums

ConnectedPoint
How we connected to a node.

NodeHandlerEndpoint
Endpoint for a received substream.

NodeHandlerEvent
Event produced by a handler.

Peer
State of a peer in the system.

RawSwarmEvent
Event that can happen on the RawSwarm.

Traits

ConnectionInfo
Information about a connection.

Type Definitions

Substream
A successfully opened substream.