Module libp2p::core::transport

Connection-oriented communication channels.

The main entity of this module is the Transport trait, which provides an interface for establishing connections with other nodes, thereby negotiating any desired protocols. The rest of the module defines combinators for modifying a transport through composition with other transports or protocol upgrades.

Modules

and_then

choice

dummy

map

map_err

memory

timeout

Transports with timeouts on the connection setup.

upgrade

Configuration of transport protocol upgrades.

Structs

Boxed

A Boxed transport is a Transport whose Dial, Listener and ListenerUpgrade futures are Boxed and only the Output and Error types are captured in type variables.

MemoryTransport

Transport that supports /memory/N multiaddresses.

OptionalTransport

Transport that is possibly disabled.

OrTransport

Struct returned by or_transport().

Upgrade

A custom upgrade on an Authenticated transport.

Enums

ListenerEvent

Event produced by Transport::Listeners.

TransportError

An error during dialing or listening on a Transport.

Traits

Transport

A transport provides connection-oriented communication between two peers through ordered streams of data (i.e. connections).