libp2p

Function tokio_development_transport

pub fn tokio_development_transport(
    keypair: Keypair,
) -> Result<Boxed<(PeerId, StreamMuxerBox)>>

👎Deprecated: Use libp2p::SwarmBuilder instead.

Builds a Transport based on TCP/IP that supports the most commonly-used features of libp2p:

DNS resolution.

Noise protocol encryption.

Websockets.

Yamux for substream multiplexing.

All async I/O of the transport is based on tokio.

Note: This Transport is not suitable for production usage, as its implementation reserves the right to support additional protocols or remove deprecated protocols.