Crate libp2p::gossipsub

Gossipsub is a P2P pubsub (publish/subscription) routing layer designed to extend upon floodsub and meshsub routing protocols.

Overview

Note: The gossipsub protocol specifications (https://github.com/libp2p/specs/tree/master/pubsub/gossipsub) provide an outline for the routing protocol. They should be consulted for further detail.

Gossipsub is a blend of meshsub for data and randomsub for mesh metadata. It provides bounded degree and amplification factor with the meshsub construction and augments it using gossip propagation of metadata with the randomsub technique.

The router maintains an overlay mesh network of peers on which to efficiently send messages and metadata. Peers use control messages to broadcast and request known messages and subscribe/unsubscribe from topics in the mesh network.

Important Discrepancies

This section outlines the current implementation’s potential discrepancies from that of other implementations, due to undefined elements in the current specification.

Topics - In gossipsub, topics configurable by the hash_topics configuration parameter. Topics are of type TopicHash. The current go implementation uses raw utf-8 strings, and this is default configuration in rust-libp2p. Topics can be hashed (SHA256 hashed then base64 encoded) by setting the hash_topics configuration parameter to true.

Sequence Numbers - A message on the gossipsub network is identified by the source libp2p_core::PeerId and a nonce (sequence number) of the message. The sequence numbers in this implementation are sent as raw bytes across the wire. They are 64-bit big-endian unsigned integers. They are chosen at random in this implementation of gossipsub, but are sequential in the current go implementation.

Using Gossipsub

GossipsubConfig

The GossipsubConfig struct specifies various network performance/tuning configuration parameters. Specifically it specifies:

This struct implements the Default trait and can be initialised via GossipsubConfig::default().

Gossipsub

The Gossipsub struct implements the libp2p_swarm::NetworkBehaviour trait allowing it to act as the routing behaviour in a libp2p_swarm::Swarm. This struct requires an instance of libp2p_core::PeerId and GossipsubConfig.

Example

An example of initialising a gossipsub compatible swarm:

use libp2p_gossipsub::GossipsubEvent;
use libp2p_core::{identity::Keypair,transport::{Transport, MemoryTransport}, Multiaddr};
use libp2p_gossipsub::MessageAuthenticity;
let local_key = Keypair::generate_ed25519();
let local_peer_id = libp2p_core::PeerId::from(local_key.public());

// Set up an encrypted TCP Transport over the Mplex
// This is test transport (memory).
let noise_keys = libp2p_noise::Keypair::<libp2p_noise::X25519Spec>::new().into_authentic(&local_key).unwrap();
let transport = MemoryTransport::default()
           .upgrade(libp2p_core::upgrade::Version::V1)
           .authenticate(libp2p_noise::NoiseConfig::xx(noise_keys).into_authenticated())
           .multiplex(libp2p_mplex::MplexConfig::new())
           .boxed();

// Create a Gossipsub topic
let topic = libp2p_gossipsub::IdentTopic::new("example");

// Set the message authenticity - How we expect to publish messages
// Here we expect the publisher to sign the message with their key.
let message_authenticity = MessageAuthenticity::Signed(local_key);

// Create a Swarm to manage peers and events
let mut swarm = {
    // set default parameters for gossipsub
    let gossipsub_config = libp2p_gossipsub::GossipsubConfig::default();
    // build a gossipsub network behaviour
    let mut gossipsub: libp2p_gossipsub::Gossipsub =
        libp2p_gossipsub::Gossipsub::new(message_authenticity, gossipsub_config).unwrap();
    // subscribe to the topic
    gossipsub.subscribe(&topic);
    // create the swarm
    libp2p_swarm::Swarm::new(
        transport,
        gossipsub,
        local_peer_id,
    )
};

// Listen on a memory transport.
let memory: Multiaddr = libp2p_core::multiaddr::Protocol::Memory(10).into();
let addr = swarm.listen_on(memory).unwrap();
println!("Listening on {:?}", addr);

Modules

error

Error types that can result from gossipsub.

protocol

subscription_filter

time_cache

This implements a time-based LRU cache for checking gossipsub message duplicates.

Structs

FastMessageId

Gossipsub

Network behaviour that handles the gossipsub protocol.

GossipsubConfig

Configuration parameters that define the performance of the gossipsub network.

GossipsubConfigBuilder

The builder struct for constructing a gossipsub configuration.

GossipsubMessage

The message sent to the user after a RawGossipsubMessage has been transformed by a crate::DataTransform.

GossipsubRpc

An RPC received/sent.

IdentityTransform

The default transform, the raw data is propagated as is to the application layer gossipsub.

MessageId

PeerScoreParams

PeerScoreThresholds

RawGossipsubMessage

A message received by the gossipsub system and stored locally in caches..

Topic

A gossipsub topic.

TopicHash

TopicScoreParams

Enums

GossipsubEvent

Event that can be emitted by the gossipsub behaviour.

MessageAcceptance

Validation kinds from the application for received messages.

MessageAuthenticity

Determines if published messages should be signed or not.

ValidationMode

The types of message validation that can be employed by gossipsub.

Traits

DataTransform

A general trait of transforming a RawGossipsubMessage into a GossipsubMessage. The RawGossipsubMessage is obtained from the wire and the GossipsubMessage is used to calculate the crate::MessageId of the message and is what is sent to the application.

Hasher

A generic trait that can be extended for various hashing types for a topic.

Functions

score_parameter_decay

Computes the decay factor for a parameter, assuming the decay_interval is 1s and that the value decays to zero if it drops below 0.01.

score_parameter_decay_with_base

Computes the decay factor for a parameter using base as the decay_interval.

Type Definitions

IdentTopic

Sha256Topic