Module libp2p::gossipsub::metrics

A set of metrics used to help track and diagnose the network behaviour of the gossipsub protocol.

Structs

Config

Metrics

A collection of metrics used throughout the Gossipsub behaviour.

Enums

Churn

Reasons why a peer was removed from the mesh.

Inclusion

Reasons why a peer was included in the mesh.

Penalty

Kinds of reasons a peer’s score has been penalized