Module libp2p::identify

Implementation of the Identify protocol.

This implementation of the protocol periodically exchanges IdentifyInfo messages between the peers on an established connection.

At least one identification request is sent on a newly established connection, beyond which the behaviour does not keep connections alive.

Usage

The Identify struct implements a NetworkBehaviour that negotiates and executes the protocol on every established connection, emitting IdentifyEvents.

Structs

Identify
Network behaviour that automatically identifies nodes periodically, returns information about them, and answers identify queries from other nodes.

IdentifyInfo
Information of a peer sent in Identify protocol responses.

Enums

IdentifyEvent
Event emitted by the Identify behaviour.