Crate libp2p::identify

Implementation of the Identify protocol.

This implementation of the protocol periodically exchanges Info messages between the peers on an established connection.

At least one identification request is sent on a newly established connection, beyond which the behaviour does not keep connections alive.

Important Discrepancies

Using Identify with other protocols Unlike some other libp2p implementations, rust-libp2p does not treat Identify as a core protocol. This means that other protocols cannot rely upon the existence of Identify, and need to be manually hooked up to Identify in order to make use of its capabilities.

Usage

The Behaviour struct implements a NetworkBehaviour that negotiates and executes the protocol on every established connection, emitting Events.

Structs

Behaviour

Network behaviour that automatically identifies nodes periodically, returns information about them, and answers identify queries from other nodes.

Config

Configuration for the identify::Behaviour.

Info

Information of a peer sent in protocol messages.

Enums

Event

Event emitted by the Identify behaviour.

UpgradeError

Constants

PROTOCOL_NAME

PUSH_PROTOCOL_NAME