Module libp2p::kad

Kademlia protocol. Allows peer discovery, records store and records fetch.

Usage

Usage is done in the following steps:

Build a KadSystemConfig and a KadConnecConfig object that contain the way you want the Kademlia protocol to behave.

Create a swarm that upgrades incoming connections with the KadConnecConfig.

Build a KadSystem from the KadSystemConfig. This requires passing a closure that provides the Kademlia controller of a peer.

You can perform queries using the KadSystem.

Modules

handler

kbucket
A k-buckets table allows one to store a value identified by keys, ordered by their distance to a reference key passed to the constructor.

protocol
Provides the KadRequestMsg and KadResponseMsg enums of all the possible messages transmitted with the Kademlia protocol, and the KademliaProtocolConfig connection upgrade.

Structs

Kademlia
Network behaviour that handles Kademlia.

Enums

KadConnectionType
Status of our connection to a node reported by the Kademlia protocol.

KademliaOut
Output event of the Kademlia behaviour.

Traits

KBucketsPeerId
Trait that must be implemented on types that can be used as an identifier in a k-bucket.