Enum libp2p::kad::handler::KademliaHandlerEvent

pub enum KademliaHandlerEvent<TUserData> {
    FindNodeReq {
        key: PeerId,
        request_id: KademliaRequestId,
    },
    FindNodeRes {
        closer_peers: Vec<KadPeer>,
        user_data: TUserData,
    },
    GetProvidersReq {
        key: Multihash,
        request_id: KademliaRequestId,
    },
    GetProvidersRes {
        closer_peers: Vec<KadPeer>,
        provider_peers: Vec<KadPeer>,
        user_data: TUserData,
    },
    QueryError {
        error: KademliaHandlerQueryErr,
        user_data: TUserData,
    },
    AddProvider {
        key: Multihash,
        provider_peer: KadPeer,
    },
    GetValue {
        key: Multihash,
        request_id: KademliaRequestId,
    },
    GetValueRes {
        result: Option<Record>,
        closer_peers: Vec<KadPeer>,
        user_data: TUserData,
    },
    PutValue {
        key: Multihash,
        value: Vec<u8>,
        request_id: KademliaRequestId,
    },
    PutValueRes {
        key: Multihash,
        user_data: TUserData,
    },
}

Event produced by the Kademlia handler.

Variants
FindNodeReq

Request for the list of nodes whose IDs are the closest to key. The number of nodes returned is not specified, but should be around 20.

Fields of FindNodeReq

key: PeerId

Identifier of the node.

request_id: KademliaRequestId

Identifier of the request. Needs to be passed back when answering.

FindNodeRes

Response to an KademliaHandlerIn::FindNodeReq.

Fields of FindNodeRes

closer_peers: Vec<KadPeer>

Results of the request.

user_data: TUserData

The user data passed to the FindNodeReq.

GetProvidersReq

Same as FindNodeReq, but should also return the entries of the local providers list for this key.

Fields of GetProvidersReq

key: Multihash

Identifier being searched.

request_id: KademliaRequestId

Identifier of the request. Needs to be passed back when answering.

GetProvidersRes

Response to an KademliaHandlerIn::GetProvidersReq.

Fields of GetProvidersRes

closer_peers: Vec<KadPeer>

Nodes closest to the key.

provider_peers: Vec<KadPeer>

Known providers for this key.

user_data: TUserData

The user data passed to the GetProvidersReq.

QueryError

An error happened when performing a query.

Fields of QueryError

error: KademliaHandlerQueryErr

The error that happened.

user_data: TUserData

The user data passed to the query.

AddProvider

The remote indicates that this list of providers is known for this key.

Fields of AddProvider

key: Multihash

Key for which we should add providers.

provider_peer: KadPeer

Known provider for this key.

GetValue

Request to get a value from the dht records

Fields of GetValue

key: Multihash

Key for which we should look in the dht

request_id: KademliaRequestId

Identifier of the request. Needs to be passed back when answering.

GetValueRes

Response to a KademliaHandlerIn::GetValue.

Fields of GetValueRes

result: Option<Record>

The result is present if the key has been found

closer_peers: Vec<KadPeer>

Nodes closest to the key.

user_data: TUserData

The user data passed to the GetValue.

PutValue

Request to put a value in the dht records

Fields of PutValue

key: Multihash

The key of the record

value: Vec<u8>

The value of the record

request_id: KademliaRequestId

Identifier of the request. Needs to be passed back when answering.

PutValueRes

Response to a request to put a value

Fields of PutValueRes

key: Multihash

The key we were putting in

user_data: TUserData

The user data passed to the GetValue.