Module libp2p::kad::protocol

👎Deprecated: The protocol module will be made private in the future and should not be depended on.

Structs

KadPeer

Information about a peer, as known by the sender.

KademliaProtocolConfig

Configuration for a Kademlia connection upgrade. When applied to a connection, turns this connection into a Stream + Sink whose items are of type KadRequestMsg and KadResponseMsg.

Enums

KadConnectionType

Status of our connection to a node reported by the Kademlia protocol.

KadRequestMsg

Request that we can send to a peer or that we received from a peer.

KadResponseMsg

Response that we can send to a peer or that we received from a peer.

Constants

DEFAULT_MAX_PACKET_SIZE

The default maximum size for a varint length-delimited packet.

DEFAULT_PROTO_NAME

The protocol name used for negotiating with multistream-select.

Type Aliases

KadInStreamSink

Sink of responses and stream of requests.

KadOutStreamSink

Sink of requests and stream of responses.

KadStreamSink