Module libp2p::kad::protocol

The Kademlia connection protocol upgrade and associated message types.

The connection protocol upgrade is provided by KademliaProtocolConfig, with the request and response types KadRequestMsg and KadResponseMsg, respectively. The upgrade’s output is a Sink + Stream of messages. The Stream component is used to poll the underlying transport for incoming messages, and the Sink component is used to send messages to remote peers.

Structs

KadPeer

Information about a peer, as known by the sender.

KademliaProtocolConfig

Configuration for a Kademlia connection upgrade. When applied to a connection, turns this connection into a Stream + Sink whose items are of type KadRequestMsg and KadResponseMsg.

Enums

KadConnectionType

Status of our connection to a node reported by the Kademlia protocol.

KadRequestMsg

Request that we can send to a peer or that we received from a peer.

KadResponseMsg

Response that we can send to a peer or that we received from a peer.

Constants

DEFAULT_MAX_PACKET_SIZE

The default maximum size for a varint length-delimited packet.

DEFAULT_PROTO_NAME

The protocol name used for negotiating with multistream-select.

Type Definitions

KadInStreamSink

Sink of responses and stream of requests.

KadOutStreamSink

Sink of requests and stream of responses.

KadStreamSink