Struct libp2p::kad::Kademlia

pub struct Kademlia<TSubstream, TRecordStorage = MemoryRecordStorage> where
    TRecordStorage: RecordStore,  { /* fields omitted */ }

Network behaviour that handles Kademlia.

Methods

impl<TSubstream, TRecordStorage> Kademlia<TSubstream, TRecordStorage> where
TRecordStorage: RecordStore,

pub fn new(local_peer_id: PeerId) -> Kademlia<TSubstream, TRecordStorage> where
TRecordStorage: Default,

Creates a Kademlia.

pub fn with_protocol_name<impl Into>>(
local_peer_id: PeerId,
name: impl Into>
) -> Kademlia<TSubstream, TRecordStorage> where
TRecordStorage: Default,
impl Into>: Into<Cow<'static, [u8]>>,

The same as new, but using a custom protocol name.

Kademlia nodes only communicate with other nodes using the same protocol name. Using a custom name therefore allows to segregate the DHT from others, if that is desired.

pub fn with_storage(
local_peer_id: PeerId,
records: TRecordStorage
) -> Kademlia<TSubstream, TRecordStorage>

The same as new, but with a custom storage.

The default records storage is in memory, this lets override the storage with user-defined behaviour

pub fn without_init(
local_peer_id: PeerId
) -> Kademlia<TSubstream, TRecordStorage> where
TRecordStorage: Default,

Deprecated:
this function is now equivalent to new() and will be removed in the future

Creates a Kademlia.

Contrary to new, doesn't perform the initialization queries that store our local ID into the DHT and fill our buckets.

pub fn add_address(&mut self, peer_id: &PeerId, address: Multiaddr)

Adds a known address of a peer participating in the Kademlia DHT to the routing table.

This allows prepopulating the Kademlia routing table with known addresses, so that they can be used immediately in following DHT operations involving one of these peers, without having to dial them upfront.

pub fn kbuckets_entries(&mut self) -> impl Iterator<Item = &PeerId>

Returns an iterator over all peer IDs of nodes currently contained in a bucket of the Kademlia routing table.

pub fn find_node(&mut self, peer_id: PeerId)

Starts an iterative FIND_NODE request.

This will eventually produce an event containing the nodes of the DHT closest to the requested PeerId.

pub fn get_providers(&mut self, target: Multihash)

Starts an iterative GET_PROVIDERS request.

pub fn get_value(&mut self, key: &Multihash, num_results: NonZeroU8)

Starts an iterative GET_VALUE request.

Returns a number of results that is in the interval [1, 20], if the user requested a larger amount of results it is cropped to 20.

pub fn put_value(&mut self, key: Multihash, value: Vec<u8>)

Starts an iterative PUT_VALUE request

pub fn add_providing(&mut self, key: Multihash)

Register the local node as the provider for the given key.

This will periodically send ADD_PROVIDER messages to the nodes closest to the key. When someone performs a GET_PROVIDERS iterative request on the DHT, our local node will be returned as part of the results.

The actual meaning of providing the value of a key is not defined, and is specific to the value whose key is the hash.

pub fn remove_providing(&mut self, key: &Multihash)

Cancels a registration done with add_providing.

There doesn't exist any "remove provider" message to broadcast on the network, therefore we will still be registered as a provider in the DHT for as long as the timeout doesn't expire.