Struct libp2p::request_response::RequestResponse

pub struct RequestResponse<TCodec> where
    TCodec: RequestResponseCodec,  { /* fields omitted */ }

A request/response protocol for some message codec.

Implementations

impl<TCodec> RequestResponse<TCodec> where
TCodec: RequestResponseCodec + Clone,

pub fn new<I>(
codec: TCodec,
protocols: I,
cfg: RequestResponseConfig
) -> RequestResponse<TCodec> where
I: IntoIterator<Item = (<TCodec as RequestResponseCodec>::Protocol, ProtocolSupport)>,

Creates a new RequestResponse behaviour for the given protocols, codec and configuration.

pub fn throttled<I>(
c: TCodec,
protos: I,
cfg: RequestResponseConfig
) -> Throttled<TCodec> where
I: IntoIterator<Item = (<TCodec as RequestResponseCodec>::Protocol, ProtocolSupport)>,
TCodec: Send,
<TCodec as RequestResponseCodec>::Protocol: Sync,

Creates a RequestResponse which limits requests per peer.

The behaviour is wrapped in Throttled and detects the limits per peer at runtime which are then enforced.

pub fn send_request(
&mut self,
peer: &PeerId,
request: <TCodec as RequestResponseCodec>::Request
) -> RequestId

Initiates sending a request.

If the targeted peer is currently not connected, a dialing attempt is initiated and the request is sent as soon as a connection is established.

Note: In order for such a dialing attempt to succeed, the RequestResonse protocol must either be embedded in another NetworkBehaviour that provides peer and address discovery, or known addresses of peers must be managed via RequestResponse::add_address and RequestResponse::remove_address.

pub fn send_response(
&mut self,
ch: ResponseChannel<<TCodec as RequestResponseCodec>::Response>,
rs: <TCodec as RequestResponseCodec>::Response
)

Initiates sending a response to an inbound request.

If the ResponseChannel is already closed due to a timeout, the response is discarded and eventually RequestResponseEvent::InboundFailure is emitted by RequestResponse::poll.

The provided ResponseChannel is obtained from a RequestResponseMessage::Request.

pub fn add_address(&mut self, peer: &PeerId, address: Multiaddr)

Adds a known address for a peer that can be used for dialing attempts by the Swarm, i.e. is returned by NetworkBehaviour::addresses_of_peer.

Addresses added in this way are only removed by remove_address.

pub fn remove_address(&mut self, peer: &PeerId, address: &Multiaddr)

Removes an address of a peer previously added via add_address.

pub fn is_connected(&self, peer: &PeerId) -> bool

Checks whether a peer is currently connected.

pub fn is_pending_outbound(&self, req_id: &RequestId) -> bool

Checks whether an outbound request initiated by RequestResponse::send_request is still pending, i.e. waiting for a response.