Crate libp2p::swarm

High level manager of the network.

A Swarm contains the state of the network as a whole. The entire behaviour of a libp2p network can be controlled through the Swarm. The Swarm struct contains all active and pending connections to remotes and manages the state of all the substreams that have been opened, and all the upgrades that were built upon these substreams.

Initializing a Swarm

Creating a Swarm requires three things:

A network identity of the local node in form of a PeerId.

An implementation of the Transport trait. This is the type that will be used in order to reach nodes on the network based on their address. See the transport module for more information.

An implementation of the NetworkBehaviour trait. This is a state machine that defines how the swarm should behave once it is connected to a node.

Network Behaviour

The NetworkBehaviour trait is implemented on types that indicate to the swarm how it should behave. This includes which protocols are supported and which nodes to try to connect to. It is the NetworkBehaviour that controls what happens on the network. Multiple types that implement NetworkBehaviour can be composed into a single behaviour.

Protocols Handler

The ConnectionHandler trait defines how each active connection to a remote should behave: how to handle incoming substreams, which protocols are supported, when to open a new outbound substream, etc.

Modules

behaviour

dial_opts

handler

Once a connection to a remote peer is established, a ConnectionHandler negotiates and handles one or more specific protocols on the connection.

Structs

AddressRecord

An record in a prioritised list of addresses.

ConnectionCounters

Network connection information.

ConnectionHandlerSelect

Implementation of ConnectionHandler that combines two protocols into one.

ConnectionLimit

Information about a connection limit.

ConnectionLimits

The configurable connection limits.

DummyBehaviour

Dummy implementation of NetworkBehaviour that doesn’t do anything.

IntoConnectionHandlerSelect

Implementation of IntoConnectionHandler that combines two protocols into one.

NetworkInfo

Information about the connections obtained by Swarm::network_info().

OneShotHandler

A ConnectionHandler that opens a new substream for each request.

OneShotHandlerConfig

Configuration parameters for the OneShotHandler

SubstreamProtocol

Configuration of inbound or outbound substream protocol(s) for a ConnectionHandler.

Swarm

Contains the state of the network, plus the way it should behave.

SwarmBuilder

A SwarmBuilder provides an API for configuring and constructing a Swarm.

SwarmPollParameters

Parameters passed to poll(), that the NetworkBehaviour has access to.

Enums

AddAddressResult

The result of adding an address to an ordered list of addresses with associated scores.

AddressScore

The “score” of an address w.r.t. an ordered collection of addresses.

CloseConnection

The options which connections to close.

ConnectionError

Errors that can occur in the context of an established Connection.

ConnectionHandlerEvent

Event produced by a handler.

ConnectionHandlerUpgrErr

Error that can happen on an outbound substream opening attempt.

DialError

The possible failures of dialing.

KeepAlive

How long the connection should be kept alive.

NetworkBehaviourAction

An action that a NetworkBehaviour can trigger in the Swarm in whose context it is executing.

NotifyHandler

The options w.r.t. which connection handler to notify of an event.

PendingConnectionError

Errors that can occur in the context of a pending Connection.

SwarmEvent

Event generated by the Swarm.

Traits

ConnectionHandler

A handler for a set of protocols used on a connection with a remote.

IntoConnectionHandler

Prototype for a ConnectionHandler.

NetworkBehaviour

A NetworkBehaviour defines the behaviour of the local node on the network.

NetworkBehaviourEventProcess

When deriving NetworkBehaviour this trait must by default be implemented for all the possible event types generated by the inner behaviours.

PollParameters

Parameters passed to poll(), that the NetworkBehaviour has access to.

Type Definitions

NegotiatedSubstream

Substream for which a protocol has been chosen.

PendingInboundConnectionError

Errors that can occur in the context of a pending incoming Connection.

PendingOutboundConnectionError

Errors that can occur in the context of a pending outgoing Connection.