libp2p::swarm

Module behaviour

Modules

toggle

Structs

AddressChange

FromSwarm variant that informs the behaviour that the ConnectedPoint of an existing connection has changed.

ConnectionClosed

FromSwarm variant that informs the behaviour about a closed connection to a peer.

ConnectionEstablished

FromSwarm variant that informs the behaviour about a newly established connection to a peer.

DialFailure

FromSwarm variant that informs the behaviour that the dial to a known or unknown node failed.

ExpiredListenAddr

FromSwarm variant that informs the behaviour that a multiaddr we were listening on has expired, which means that we are no longer listening on it.

ExternalAddrConfirmed

FromSwarm variant that informs the behaviour that an external address was confirmed.

ExternalAddrExpired

FromSwarm variant that informs the behaviour that an external address was removed.

ExternalAddresses

Utility struct for tracking the external addresses of a Swarm.

ListenAddresses

Utility struct for tracking the addresses a Swarm is listening on.

ListenFailure

FromSwarm variant that informs the behaviour that an error happened on an incoming connection during its initial handshake.

ListenerClosed

FromSwarm variant that informs the behaviour that a listener closed.

ListenerError

FromSwarm variant that informs the behaviour that a listener experienced an error.

NewExternalAddrCandidate

FromSwarm variant that informs the behaviour about a new candidate for an external address for us.

NewListenAddr

FromSwarm variant that informs the behaviour that we have started listening on a new multiaddr.

NewListener

FromSwarm variant that informs the behaviour that a new listener was created.

Enums

CloseConnection

The options which connections to close.

FromSwarm

Enumeration with the list of the possible events to pass to on_swarm_event.

NotifyHandler

The options w.r.t. which connection handler to notify of an event.

ToSwarm

A command issued from a NetworkBehaviour for the Swarm.

Traits

NetworkBehaviour

A NetworkBehaviour defines the behaviour of the local node on the network.

PollParameters

Parameters passed to poll(), that the NetworkBehaviour has access to.