Module libp2p::swarm::behaviour

Modules

toggle

Enums

CloseConnection

The options which connections to close.

NetworkBehaviourAction

An action that a NetworkBehaviour can trigger in the Swarm in whose context it is executing.

NotifyHandler

The options w.r.t. which connection handler to notify of an event.

Traits

NetworkBehaviour

A NetworkBehaviour defines the behaviour of the local node on the network.

NetworkBehaviourEventProcess

When deriving NetworkBehaviour this trait must by default be implemented for all the possible event types generated by the inner behaviours.

PollParameters

Parameters passed to poll(), that the NetworkBehaviour has access to.