Module libp2p::swarm::handler

Once a connection to a remote peer is established, a ConnectionHandler negotiates and handles one or more specific protocols on the connection.

Protocols are negotiated and used on individual substreams of the connection. Thus a ConnectionHandler defines the inbound and outbound upgrades to apply when creating a new inbound or outbound substream, respectively, and is notified by a Swarm when these upgrades have been successfully applied, including the final output of the upgrade. A ConnectionHandler can then continue communicating with the peer over the substream using the negotiated protocol(s).

Two ConnectionHandlers can be composed with ConnectionHandler::select() in order to build a new handler supporting the combined set of protocols, with methods being dispatched to the appropriate handler according to the used protocol(s) determined by the associated types of the handlers.

Note: A ConnectionHandler handles one or more protocols in the context of a single connection with a remote. In order to handle a protocol that requires knowledge of the network as a whole, see the NetworkBehaviour trait.

Modules

either

multi

A ConnectionHandler implementation that combines multiple other ConnectionHandlers indexed by some key.

Structs

ConnectionHandlerSelect

Implementation of ConnectionHandler that combines two protocols into one.

DummyConnectionHandler

Implementation of ConnectionHandler that doesn’t handle anything.

IntoConnectionHandlerSelect

Implementation of IntoConnectionHandler that combines two protocols into one.

MapInEvent

Wrapper around a protocol handler that turns the input event into something else.

MapOutEvent

Wrapper around a protocol handler that turns the output event into something else.

OneShotHandler

A ConnectionHandler that opens a new substream for each request.

OneShotHandlerConfig

Configuration parameters for the OneShotHandler

SendWrapper

Wraps around a type that implements OutboundUpgradeSend, InboundUpgradeSend, or both, and implements OutboundUpgrade and/or InboundUpgrade.

SubstreamProtocol

Configuration of inbound or outbound substream protocol(s) for a ConnectionHandler.

Enums

ConnectionHandlerEvent

Event produced by a handler.

ConnectionHandlerUpgrErr

Error that can happen on an outbound substream opening attempt.

KeepAlive

How long the connection should be kept alive.

Traits

ConnectionHandler

A handler for a set of protocols used on a connection with a remote.

InboundUpgradeSend

Implemented automatically on all types that implement InboundUpgrade and Send + 'static.

IntoConnectionHandler

Prototype for a ConnectionHandler.

OutboundUpgradeSend

Implemented automatically on all types that implement OutboundUpgrade and Send + 'static.

UpgradeInfoSend

Implemented automatically on all types that implement UpgradeInfo and Send + 'static.