Trait libp2p::swarm::handler::IntoConnectionHandler

pub trait IntoConnectionHandler: Send + 'static {
    type Handler: ConnectionHandler;

    // Required methods
    fn into_handler(
        self,
        remote_peer_id: &PeerId,
        connected_point: &ConnectedPoint,
    ) -> Self::Handler;
    fn inbound_protocol(
        &self,
    ) -> <Self::Handler as ConnectionHandler>::InboundProtocol;

    // Provided method
    fn select<TProto2>(
        self,
        other: TProto2,
    ) -> IntoConnectionHandlerSelect<Self, TProto2>
       where Self: Sized { ... }
}

👎Deprecated: Implement ConnectionHandler directly and use NetworkBehaviour::{handle_pending_inbound_connection,handle_pending_outbound_connection} to handle pending connections.

Prototype for a ConnectionHandler.

Required Associated Types

type Handler: ConnectionHandler

👎Deprecated: Implement ConnectionHandler directly and use NetworkBehaviour::{handle_pending_inbound_connection,handle_pending_outbound_connection} to handle pending connections.

The protocols handler.

Required Methods

fn into_handler( self, remote_peer_id: &PeerId, connected_point: &ConnectedPoint, ) -> Self::Handler

👎Deprecated: Implement ConnectionHandler directly and use NetworkBehaviour::{handle_pending_inbound_connection,handle_pending_outbound_connection} to handle pending connections.

Builds the protocols handler.

The PeerId is the id of the node the handler is going to handle.

fn inbound_protocol( &self, ) -> <Self::Handler as ConnectionHandler>::InboundProtocol

👎Deprecated: Implement ConnectionHandler directly and use NetworkBehaviour::{handle_pending_inbound_connection,handle_pending_outbound_connection} to handle pending connections.

Return the handler’s inbound protocol.

Provided Methods

fn select<TProto2>( self, other: TProto2, ) -> IntoConnectionHandlerSelect<Self, TProto2>
where Self: Sized,

👎Deprecated: Implement ConnectionHandler directly and use NetworkBehaviour::{handle_pending_inbound_connection,handle_pending_outbound_connection} to handle pending connections.

Builds an implementation of IntoConnectionHandler that handles both this protocol and the other one together.