Crate libp2p::tcp

Implementation of the libp2p Transport trait for TCP/IP.

Usage

This crate provides a TcpTransport and TokioTcpTransport, depending on the enabled features, which implement the Transport trait for use as a transport with libp2p-core or libp2p-swarm.

Modules

async_io

Structs

GenTcpConfig

The configuration for a TCP/IP transport capability for libp2p.

GenTcpTransport

TcpListenStream

A stream of incoming connections on one or more interfaces.

Enums

TcpListenerEvent

Event produced by a TcpListenStream.

Type Definitions

TcpTransport

The type of a GenTcpTransport using the async-io implementation.