Crate libp2p::tcp

Implementation of the libp2p Transport trait for TCP/IP.

Usage

This crate provides a TcpConfig and TokioTcpConfig, depending on the enabled features, which implement the Transport trait for use as a transport with libp2p-core or libp2p-swarm.

Modules

async_io

Structs

GenTcpConfig

The configuration for a TCP/IP transport capability for libp2p.

TcpListenStream

A stream of incoming connections on one or more interfaces.

Type Definitions

TcpConfig

The type of a GenTcpConfig using the async-io implementation.