Trait libp2p::TransportExt

pub trait TransportExt: Transport {
    fn with_bandwidth_logging(
        self
    ) -> (BandwidthLogging<Self>, Arc<BandwidthSinks>)
    where
        Self: Sized,
    { ... }
}

Trait automatically implemented on all objects that implement Transport. Provides some additional utilities.

Provided Methods

fn with_bandwidth_logging(self) -> (BandwidthLogging<Self>, Arc<BandwidthSinks>)where
Self: Sized,

Adds a layer on the Transport that logs all trafic that passes through the sockets created by it.

This method returns an Arc<BandwidthSinks> that can be used to retreive the total number of bytes transferred through the sockets.