Module libp2p::uds

Implementation of the libp2p Transport trait for Unix domain sockets.

Uses the tokio library.

Platform support

This transport only works on Unix platforms.

Usage

The UdsConfig transport supports multiaddresses of the form /unix//tmp/foo.

Example:

extern crate libp2p_uds;
use libp2p_uds::UdsConfig;

let uds = UdsConfig::new();

The UdsConfig structs implements the Transport trait of the core library. See the documentation of core and of libp2p in general to learn how to use the Transport trait.

Structs

ListenerStream

UdsConfig
Represents the configuration for a Unix domain sockets transport capability for libp2p.