Crate libp2p::yamux

Implements the Yamux multiplexing protocol for libp2p, see also the specification.

Structs

Incoming
The futures::stream::Stream of incoming substreams.

LocalIncoming
The futures::stream::Stream of incoming substreams (!Send).

OpenSubstreamToken
A token to poll for an outbound substream.

WindowUpdateMode
The window update mode determines when window updates are sent to the remote, giving it new credit to send more data.

Yamux
A Yamux connection.

YamuxConfig
The yamux configuration.

YamuxError
The Yamux StreamMuxer error type.

YamuxLocalConfig
The yamux configuration for upgrading I/O resources which are
Send
.

Type Definitions

YamuxResult