Crate libp2p::yamux

Implementation of the Yamux multiplexing protocol for libp2p.

Structs

Config

The yamux configuration.

Error

The Yamux StreamMuxer error type.

Incoming

The futures::stream::Stream of incoming substreams.

LocalConfig

The yamux configuration for upgrading I/O resources which are ![Send].

LocalIncoming

The futures::stream::Stream of incoming substreams (!Send).

Muxer

A Yamux connection.

WindowUpdateMode

The window update mode determines when window updates are sent to the remote, giving it new credit to send more data.

Type Aliases

YamuxDeprecated

YamuxConfigDeprecated

YamuxErrorDeprecated

YamuxLocalConfigDeprecated

YamuxResultDeprecated